Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3         4         5        6        7
#> 1   0.000000 7.3589504 2.7244425 7.0909893 3.7122690 6.000933 6.284331
#> 2   7.358950 0.0000000 4.6677372 0.6439215 3.9652304 2.033900 1.668971
#> 3   2.724442 4.6677372 0.0000000 4.4585116 1.1331430 3.303390 3.785388
#> 4   7.090989 0.6439215 4.4585116 0.0000000 3.8953330 2.286875 1.077121
#> 5   3.712269 3.9652304 1.1331430 3.8953330 0.0000000 2.312850 3.470156
#> 6   6.000933 2.0338997 3.3033895 2.2868748 2.3128505 0.000000 2.554423
#> 7   6.284331 1.6689709 3.7853876 1.0771206 3.4701560 2.554423 0.000000
#> 8   6.228750 1.3251231 3.6258585 0.8736211 3.1632383 2.024691 0.539683
#> 9   1.847159 8.1360250 3.5807368 7.9995794 4.1858386 6.443456 7.366101
#> 10  5.430101 2.9672421 3.2715959 2.3885960 3.3703999 3.351629 1.311475
#> 11 10.621340 3.3703806 7.9847398 3.5326509 7.3349635 5.290098 4.402016
#> 12  2.289014 5.0786390 0.4542469 4.8415577 1.5684880 3.756538 4.114982
#> 13  3.565276 3.8945966 1.2467119 3.5553272 1.7048474 3.052697 2.720093
#> 14  2.627281 4.7647491 0.6788834 4.4682132 1.7103607 3.654791 3.665301
#> 15 11.765560 4.8156306 9.2342302 4.8184322 8.7116465 6.821126 5.484217
#> 16  5.180619 2.5331045 2.7555300 2.0632780 2.6511273 2.558388 1.110646
#> 17  2.335621 5.0279758 0.5904503 4.7568266 1.7185664 3.811808 3.982845
#> 18  1.485674 5.9411425 1.2737055 5.7198507 2.2275243 4.517109 4.996560
#> 19  4.567539 3.3072315 1.9524486 3.3448372 0.8559186 1.491539 3.137387
#> 20  2.703177 9.7101616 5.0751920 9.5325904 5.7941134 8.077275 8.827622
#> 21  3.318404 4.0850437 0.5940091 3.8983550 0.7357191 2.724329 3.283654
#> 22  2.681250 4.7500265 0.1728100 4.5602204 1.0907226 3.323991 3.913369
#> 23  2.064698 5.5789439 1.5073980 5.2239834 2.6216154 4.589383 4.327403
#> 24  1.437188 8.6728076 4.0050719 8.4450863 4.8603040 7.172854 7.679897
#> 25  1.371388 6.2452696 1.8824027 5.9127122 3.0101523 5.134520 5.035507
#>            8         9        10        11        12        13        14
#> 1  6.2287496  1.847159 5.4301013 10.621340 2.2890144 3.5652758 2.6272807
#> 2  1.3251231  8.136025 2.9672421  3.370381 5.0786390 3.8945966 4.7647491
#> 3  3.6258585  3.580737 3.2715959  7.984740 0.4542469 1.2467119 0.6788834
#> 4  0.8736211  7.999579 2.3885960  3.532651 4.8415577 3.5553272 4.4682132
#> 5  3.1632383  4.185839 3.3703999  7.334963 1.5684880 1.7048474 1.7103607
#> 6  2.0246914  6.443456 3.3516286  5.290098 3.7565382 3.0526971 3.6547909
#> 7  0.5396830  7.366101 1.3114753  4.402016 4.1149818 2.7200929 3.6653013
#> 8  0.0000000  7.190379 1.6878816  4.392904 3.9944120 2.6837895 3.6029258
#> 9  7.1903795  0.000000 6.7480890 11.496949 3.2784514 4.7309444 3.8132716
#> 10 1.6878816  6.748089 0.0000000  5.559738 3.4842833 2.0367835 2.9349732
#> 11 4.3929043 11.496949 5.5597378  0.000000 8.3733979 7.0717084 7.9958221
#> 12 3.9944120  3.278451 3.4842833  8.373398 0.0000000 1.4542063 0.5950903
#> 13 2.6837895  4.730944 2.0367835  7.071708 1.4542063 0.0000000 0.9478862
#> 14 3.6029258  3.813272 2.9349732  7.995822 0.5950903 0.9478862 0.0000000
#> 15 5.6118290 12.802030 6.4469660  1.706711 9.5877779 8.2037702 9.1495150
#> 16 1.2130307  6.322307 0.8272084  5.505984 3.0495090 1.6234012 2.5712832
#> 17 3.8978132  3.468202 3.2799043  8.288567 0.2915723 1.2774643 0.3463427
#> 18 4.8759891  2.429169 4.3195306  9.250224 0.8831053 2.3132459 1.3845826
#> 19 2.7314696  4.953969 3.3669896  6.664650 2.4010349 2.1621046 2.4487907
#> 20 8.6997446  1.659296 8.0805874 13.053109 4.7142935 6.1280250 5.1821736
#> 21 3.0826517  4.107802 2.9219710  7.415793 1.0383272 1.0372120 1.0083379
#> 22 3.7368670  3.457176 3.4309406  8.080452 0.4828662 1.4127915 0.8309959
#> 23 4.3503836  3.609702 3.3809242  8.718962 1.1559444 1.6845770 0.9454923
#> 24 7.5935785  1.322927 6.8619836 11.977331 3.6041083 4.9615295 4.0147252
#> 25 5.0400359  3.037728 4.0980083  9.419631 1.4423313 2.3586822 1.5034104
#>           15        16        17         18        19        20        21
#> 1  11.765560 5.1806194 2.3356212  1.4856739 4.5675385  2.703177 3.3184044
#> 2   4.815631 2.5331045 5.0279758  5.9411425 3.3072315  9.710162 4.0850437
#> 3   9.234230 2.7555300 0.5904503  1.2737055 1.9524486  5.075192 0.5940091
#> 4   4.818432 2.0632780 4.7568266  5.7198507 3.3448372  9.532590 3.8983550
#> 5   8.711646 2.6511273 1.7185664  2.2275243 0.8559186  5.794113 0.7357191
#> 6   6.821126 2.5583875 3.8118077  4.5171094 1.4915385  8.077275 2.7243292
#> 7   5.484217 1.1106465 3.9828454  4.9965598 3.1373872  8.827622 3.2836538
#> 8   5.611829 1.2130307 3.8978132  4.8759891 2.7314696  8.699745 3.0826517
#> 9  12.802030 6.3223068 3.4682019  2.4291692 4.9539688  1.659296 4.1078017
#> 10  6.446966 0.8272084 3.2799043  4.3195306 3.3669896  8.080587 2.9219710
#> 11  1.706711 5.5059841 8.2885674  9.2502240 6.6646501 13.053109 7.4157932
#> 12  9.587778 3.0495090 0.2915723  0.8831053 2.4010349  4.714293 1.0383272
#> 13  8.203770 1.6234012 1.2774643  2.3132459 2.1621046  6.128025 1.0372120
#> 14  9.149515 2.5712832 0.3463427  1.3845826 2.4487907  5.182174 1.0083379
#> 15  0.000000 6.5849427 9.4657001 10.4708540 8.1128296 14.301586 8.6942299
#> 16  6.584943 0.0000000 2.8972538  3.9245685 2.5610135  7.748797 2.3108739
#> 17  9.465700 2.8972538 0.0000000  1.0406755 2.5186602  4.851636 1.0952768
#> 18 10.470854 3.9245685 1.0406755  0.0000000 3.0832507  3.831556 1.8626468
#> 19  8.112830 2.5610135 2.5186602  3.0832507 0.0000000  6.585767 1.4421166
#> 20 14.301586 7.7487970 4.8516364  3.8315563 6.5857672  0.000000 5.6374265
#> 21  8.694230 2.3108739 1.0952768  1.8626468 1.4421166  5.637426 0.0000000
#> 22  9.348245 2.8976092 0.6880402  1.2076787 1.9331504  4.973312 0.6654815
#> 23  9.776071 3.2172181 0.9201295  1.3733455 3.3890445  4.755547 1.9469450
#> 24 13.164085 6.5839897 3.7008009  2.7319803 5.6940743  1.297874 4.5903523
#> 25 10.491774 3.9248612 1.3260852  1.1671856 3.8339578  4.074039 2.4212644
#>           22        23        24         25
#> 1  2.6812495 2.0646980  1.437188  1.3713881
#> 2  4.7500265 5.5789439  8.672808  6.2452696
#> 3  0.1728100 1.5073980  4.005072  1.8824027
#> 4  4.5602204 5.2239834  8.445086  5.9127122
#> 5  1.0907226 2.6216154  4.860304  3.0101523
#> 6  3.3239905 4.5893832  7.172854  5.1345200
#> 7  3.9133693 4.3274032  7.679897  5.0355073
#> 8  3.7368670 4.3503836  7.593578  5.0400359
#> 9  3.4571756 3.6097020  1.322927  3.0377281
#> 10 3.4309406 3.3809242  6.861984  4.0980083
#> 11 8.0804519 8.7189621 11.977331  9.4196309
#> 12 0.4828662 1.1559444  3.604108  1.4423313
#> 13 1.4127915 1.6845770  4.961529  2.3586822
#> 14 0.8309959 0.9454923  4.014725  1.5034104
#> 15 9.3482452 9.7760706 13.164085 10.4917743
#> 16 2.8976092 3.2172181  6.583990  3.9248612
#> 17 0.6880402 0.9201295  3.700801  1.3260852
#> 18 1.2076787 1.3733455  2.731980  1.1671856
#> 19 1.9331504 3.3890445  5.694074  3.8339578
#> 20 4.9733120 4.7555472  1.297874  4.0740390
#> 21 0.6654815 1.9469450  4.590352  2.4212644
#> 22 0.0000000 1.6065336  3.928068  1.9240688
#> 23 1.6065336 0.0000000  3.501886  0.7180738
#> 24 3.9280678 3.5018860  0.000000  2.8042395
#> 25 1.9240688 0.7180738  2.804240  0.0000000