Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2          3         4         5         6         7
#> 1   0.000000 5.3124095  2.5745905 1.9681262 5.1353793 2.0015981 2.3563147
#> 2   5.312410 0.0000000  7.5102687 3.5329609 0.2283778 3.3344079 3.1513257
#> 3   2.574590 7.5102687  0.0000000 3.9784221 7.3673654 4.4119503 4.3597598
#> 4   1.968126 3.5329609  3.9784221 0.0000000 3.3892008 0.9357774 0.3997185
#> 5   5.135379 0.2283778  7.3673654 3.3892008 0.0000000 3.1481935 3.0147056
#> 6   2.001598 3.3344079  4.4119503 0.9357774 3.1481935 0.0000000 0.9949699
#> 7   2.356315 3.1513257  4.3597598 0.3997185 3.0147056 0.9949699 0.0000000
#> 8   8.135375 2.8880562 10.3955759 6.4171664 3.0293981 6.1348725 6.0385112
#> 9   4.008034 1.6124147  6.4193528 2.5632648 1.3865406 2.0205757 2.2768882
#> 10  4.759774 0.5714732  6.9390873 2.9623058 0.4682561 2.7969226 2.5799149
#> 11  7.847716 2.7389257 10.1796876 6.2107439 2.8406180 5.8481773 5.8459160
#> 12  8.523487 3.4093084 10.8601754 6.8908800 3.5171961 6.5250205 6.5252708
#> 13  6.091455 1.1302636  8.4241328 4.4627990 1.1566453 4.0906164 4.1059238
#> 14  2.863715 7.2984718  0.9141197 3.8304859 7.1782369 4.4344683 4.1763417
#> 15  4.509587 0.8307740  6.6798231 2.7032763 0.7136155 2.5565012 2.3205991
#> 16  6.007551 0.9494705  8.0471585 4.1221864 1.1760904 4.0883473 3.7242287
#> 17  7.726626 2.4998084  9.9997189 6.0216272 2.6324265 5.7255810 5.6457653
#> 18 11.134420 5.9076918 13.4159362 9.4375160 6.0491671 9.1329193 9.0586976
#> 19  6.605064 1.9213628  9.0150729 5.0955507 1.9214705 4.6237469 4.7591251
#> 20  6.265065 1.0179989  8.5137800 4.5354578 1.1465922 4.2686335 4.1591632
#> 21  3.478977 8.4207919  0.9263744 4.8926333 8.2815401 5.3381978 5.2695087
#> 22  1.986574 3.3921179  4.1446465 0.2936715 3.2354577 0.6498869 0.4328839
#> 23  2.938133 2.6286237  5.4155909 1.7983656 2.4152552 1.0182488 1.6518755
#> 24  9.053769 4.0429095 11.4341278 7.4797666 4.1344383 7.0649604 7.1229382
#> 25  6.043306 0.7522817  8.2619314 4.2840953 0.9081746 4.0538231 3.9034145
#>             8        9        10         11         12        13         14
#> 1   8.1353749 4.008034 4.7597741  7.8477156  8.5234872 6.0914555  2.8637155
#> 2   2.8880562 1.612415 0.5714732  2.7389257  3.4093084 1.1302636  7.2984718
#> 3  10.3955759 6.419353 6.9390873 10.1796876 10.8601754 8.4241328  0.9141197
#> 4   6.4171664 2.563265 2.9623058  6.2107439  6.8908800 4.4627990  3.8304859
#> 5   3.0293981 1.386541 0.4682561  2.8406180  3.5171961 1.1566453  7.1782369
#> 6   6.1348725 2.020576 2.7969226  5.8481773  6.5250205 4.0906164  4.4344683
#> 7   6.0385112 2.276888 2.5799149  5.8459160  6.5252708 4.1059238  4.1763417
#> 8   0.0000000 4.177429 3.4594279  0.6139544  0.7541737 2.0755720 10.1816258
#> 9   4.1774287 0.000000 1.2693945  3.8480069  4.5201731 2.1056066  6.3727792
#> 10  3.4594279 1.269395 0.0000000  3.2976218  3.9715383 1.6245442  6.7294778
#> 11  0.6139544 3.848007 3.2976218  0.0000000  0.6804879 1.7600037 10.0169594
#> 12  0.7541737 4.520173 3.9715383  0.6804879  0.0000000 2.4394445 10.6949554
#> 13  2.0755720 2.105607 1.6245442  1.7600037  2.4394445 0.0000000  8.2821181
#> 14 10.1816258 6.372779 6.7294778 10.0169594 10.6949554 8.2821181  0.0000000
#> 15  3.7186314 1.170633 0.2593170  3.5518395  4.2269125 1.8621665  6.4718060
#> 16  2.5730924 2.539232 1.3290458  2.6333826  3.2343251 1.5623922  7.7421663
#> 17  0.4142586 3.763760 3.0699944  0.4780790  0.9747082 1.6613300  9.7980689
#> 18  3.0203907 7.144130 6.4791602  3.2988293  2.6398614 5.0442148 13.1943182
#> 19  1.8743476 2.603185 2.3821404  1.3692226  1.9891977 0.7987445  8.9260329
#> 20  1.8829957 2.368735 1.5853990  1.7265693  2.3927989 0.5345762  8.3148755
#> 21 11.3078879 7.345473 7.8493450 11.0993120 11.7797688 9.3452156  1.3631314
#> 22  6.2639906 2.327230 2.8258704  6.0351558  6.7156372 4.2797332  4.0469376
#> 23  5.2825745 1.105158 2.1675854  4.9465751  5.6154332 3.2102776  5.4525692
#> 24  1.4351765 5.046608 4.5961214  1.3100547  0.6937050 3.0173828 11.2987185
#> 25  2.1358138 2.208339 1.3237321  2.0116984  2.6730453 0.6847829  8.0486993
#>           15        16         17        18        19        20         21
#> 1  4.5095866 6.0075506  7.7266258 11.134420 6.6050643 6.2650650  3.4789772
#> 2  0.8307740 0.9494705  2.4998084  5.907692 1.9213628 1.0179989  8.4207919
#> 3  6.6798231 8.0471585  9.9997189 13.415936 9.0150729 8.5137800  0.9263744
#> 4  2.7032763 4.1221864  6.0216272  9.437516 5.0955507 4.5354578  4.8926333
#> 5  0.7136155 1.1760904  2.6324265  6.049167 1.9214705 1.1465922  8.2815401
#> 6  2.5565012 4.0883473  5.7255810  9.132919 4.6237469 4.2686335  5.3381978
#> 7  2.3205991 3.7242287  5.6457653  9.058698 4.7591251 4.1591632  5.2695087
#> 8  3.7186314 2.5730924  0.4142586  3.020391 1.8743476 1.8829957 11.3078879
#> 9  1.1706325 2.5392322  3.7637604  7.144130 2.6031845 2.3687353  7.3454726
#> 10 0.2593170 1.3290458  3.0699944  6.479160 2.3821404 1.5853990  7.8493450
#> 11 3.5518395 2.6333826  0.4780790  3.298829 1.3692226 1.7265693 11.0993120
#> 12 4.2269125 3.2343251  0.9747082  2.639861 1.9891977 2.3927989 11.7797688
#> 13 1.8621665 1.5623922  1.6613300  5.044215 0.7987445 0.5345762  9.3452156
#> 14 6.4718060 7.7421663  9.7980689 13.194318 8.9260329 8.3148755  1.3631314
#> 15 0.0000000 1.5427117  3.3286564  6.738440 2.6040223 1.8433911  7.5900297
#> 16 1.5427117 0.0000000  2.2602647  5.502670 2.3039375 1.1131701  8.9370717
#> 17 3.3286564 2.2602647  0.0000000  3.418518 1.4987914 1.4866073 10.9138563
#> 18 6.7384396 5.5026702  3.4185182  0.000000 4.6266468 4.9025749 14.3281657
#> 19 2.6040223 2.3039375  1.4987914  4.626647 0.0000000 1.1951096  9.9403560
#> 20 1.8433911 1.1131701  1.4866073  4.902575 1.1951096 0.0000000  9.4274013
#> 21 7.5900297 8.9370717 10.9138563 14.328166 9.9403560 9.4274013  0.0000000
#> 22 2.5691458 4.0333001  5.8637813  9.282240 4.8888876 4.3810908  5.0655339
#> 23 1.9750257 3.4960607  4.8689188  8.244651 3.6725131 3.4627488  6.3408389
#> 24 4.8477493 3.9190913  1.6640623  2.272003 2.4531945 3.0349096 12.3569443
#> 25 1.5830049 0.8795189  1.7496311  5.155435 1.4334640 0.2916156  9.1729200
#>           22       23        24        25
#> 1  1.9865737 2.938133  9.053769 6.0433059
#> 2  3.3921179 2.628624  4.042910 0.7522817
#> 3  4.1446465 5.415591 11.434128 8.2619314
#> 4  0.2936715 1.798366  7.479767 4.2840953
#> 5  3.2354577 2.415255  4.134438 0.9081746
#> 6  0.6498869 1.018249  7.064960 4.0538231
#> 7  0.4328839 1.651875  7.122938 3.9034145
#> 8  6.2639906 5.282575  1.435177 2.1358138
#> 9  2.3272297 1.105158  5.046608 2.2083394
#> 10 2.8258704 2.167585  4.596121 1.3237321
#> 11 6.0351558 4.946575  1.310055 2.0116984
#> 12 6.7156372 5.615433  0.693705 2.6730453
#> 13 4.2797332 3.210278  3.017383 0.6847829
#> 14 4.0469376 5.452569 11.298718 8.0486993
#> 15 2.5691458 1.975026  4.847749 1.5830049
#> 16 4.0333001 3.496061  3.919091 0.8795189
#> 17 5.8637813 4.868919  1.664062 1.7496311
#> 18 9.2822398 8.244651  2.272003 5.1554347
#> 19 4.8888876 3.672513  2.453195 1.4334640
#> 20 4.3810908 3.462749  3.034910 0.2916156
#> 21 5.0655339 6.340839 12.356944 9.1729200
#> 22 0.0000000 1.513800  7.292268 4.1387698
#> 23 1.5137998 0.000000  6.124852 3.2849760
#> 24 7.2922675 6.124852  0.000000 3.3212664
#> 25 4.1387698 3.284976  3.321266 0.0000000