Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1          2        3         4          5          6         7
#> 1  0.0000000  5.3634893 2.822148 1.6333479  3.3752985  2.8750609  5.553817
#> 2  5.3634893  0.0000000 4.276970 5.8994247  8.3599281  7.9498406 10.517166
#> 3  2.8221476  4.2769699 0.000000 2.1299263  4.5011404  4.2195934  6.504045
#> 4  1.6333479  5.8994247 2.129926 0.0000000  2.4786276  2.1269027  4.617753
#> 5  3.3752985  8.3599281 4.501140 2.4786276  0.0000000  0.5236944  2.196199
#> 6  2.8750609  7.9498406 4.219593 2.1269027  0.5236944  0.0000000  2.679382
#> 7  5.5538170 10.5171661 6.504045 4.6177531  2.1961986  2.6793818  0.000000
#> 8  1.3744341  6.6975208 4.007356 2.2158713  2.7143772  2.1956803  4.718892
#> 9  1.1987584  5.8221633 2.333490 0.4556458  2.5488463  2.1280588  4.734448
#> 10 1.5084391  3.9697746 2.749478 2.7251177  4.8450235  4.3589166  7.036497
#> 11 1.6573178  4.2865918 1.225790 1.6139001  4.0775440  3.6865412  6.231185
#> 12 5.7689926  0.5252862 4.795611 6.3777153  8.8275964  8.4080487 10.994062
#> 13 1.3730261  4.1581891 1.726767 1.8664361  4.2374859  3.8039826  6.423201
#> 14 1.2697054  4.8627396 1.574183 1.0657913  3.4972631  3.0953720  5.664623
#> 15 4.7997937  0.7775171 3.511223 5.2040032  7.6770708  7.2835995  9.816465
#> 16 3.8596772  2.1024588 2.185862 3.9898847  6.4674416  6.1047494  8.571199
#> 17 1.4663869  4.5147499 1.365222 1.3942157  3.8457779  3.4486546  6.006688
#> 18 3.1100908  6.1002174 1.825115 1.5648865  3.1182232  3.0063293  4.874422
#> 19 3.2797142  8.6008881 5.175693 3.0458331  1.3367362  1.1537885  2.692126
#> 20 2.4652096  7.1665943 3.273587 1.2673374  1.2317494  0.9860947  3.350799
#> 21 9.3308434  3.9739374 7.849694 9.7274190 12.2060365 11.8284292 14.312736
#> 22 0.7824482  5.7407508 2.560176 0.8952548  2.7256333  2.2576908  4.921608
#> 23 0.2342420  5.5977165 2.972655 1.6075522  3.1874143  2.6807292  5.356123
#> 24 2.1540957  3.3994700 2.955978 3.2913648  5.4793162  4.9976088  7.673088
#> 25 0.7606689  4.9089343 2.062473 1.2986306  3.5268342  3.0755125  5.721103
#>            8         9        10        11         12        13        14
#> 1   1.374434 1.1987584 1.5084391 1.6573178  5.7689926 1.3730261 1.2697054
#> 2   6.697521 5.8221633 3.9697746 4.2865918  0.5252862 4.1581891 4.8627396
#> 3   4.007356 2.3334900 2.7494783 1.2257902  4.7956114 1.7267668 1.5741828
#> 4   2.215871 0.4556458 2.7251177 1.6139001  6.3777153 1.8664361 1.0657913
#> 5   2.714377 2.5488463 4.8450235 4.0775440  8.8275964 4.2374859 3.4972631
#> 6   2.195680 2.1280588 4.3589166 3.6865412  8.4080487 3.8039826 3.0953720
#> 7   4.718892 4.7344478 7.0364972 6.2311846 10.9940622 6.4232007 5.6646226
#> 8   0.000000 1.8063858 2.7514666 2.9502256  7.0824677 2.7384511 2.4451470
#> 9   1.806386 0.0000000 2.4148182 1.5694868  6.2832193 1.6892345 0.9730060
#> 10  2.751467 2.4148182 0.0000000 1.6242478  4.3364169 1.0879164 1.7868056
#> 11  2.950226 1.5694868 1.6242478 0.0000000  4.7638361 0.5392322 0.5973140
#> 12  7.082468 6.2832193 4.3364169 4.7638361  0.0000000 4.6065565 5.3318465
#> 13  2.738451 1.6892345 1.0879164 0.5392322  4.6065565 0.0000000 0.8115672
#> 14  2.445147 0.9730060 1.7868056 0.5973140  5.3318465 0.8115672 0.0000000
#> 15  6.159415 5.1575400 3.4948290 3.6003657  1.3027844 3.5281738 4.1882815
#> 16  5.232733 4.0035793 2.8225296 2.4362291  2.6258701 2.4994639 3.0333891
#> 17  2.728418 1.3276880 1.6451171 0.2443488  4.9874925 0.5752938 0.3546844
#> 18  3.763176 2.0127115 3.8342909 2.2802015  6.6172598 2.7717197 2.0689490
#> 19  2.152492 2.9078850 4.7793410 4.4594921  9.0270428 4.4581518 3.8643942
#> 20  2.304931 1.4256949 3.8272241 2.8803877  7.6435488 3.0941738 2.3178071
#> 21 10.670921 9.7072897 7.9413105 8.1420120  3.6256898 8.0850702 8.7352173
#> 22  1.463397 0.4405413 2.1305372 1.6109808  6.1825565 1.5844308 1.0394577
#> 23  1.149694 1.1561888 1.7306128 1.8388480  6.0026590 1.5895570 1.4024954
#> 24  3.377086 3.0147583 0.6456859 1.9902656  3.7384478 1.5033453 2.2894661
#> 25  2.047797 1.0074324 1.4272679 0.9089880  5.3505629 0.7563965 0.5393182
#>           15       16        17       18        19         20        21
#> 1  4.7997937 3.859677 1.4663869 3.110091  3.279714  2.4652096  9.330843
#> 2  0.7775171 2.102459 4.5147499 6.100217  8.600888  7.1665943  3.973937
#> 3  3.5112231 2.185862 1.3652224 1.825115  5.175693  3.2735875  7.849694
#> 4  5.2040032 3.989885 1.3942157 1.564886  3.045833  1.2673374  9.727419
#> 5  7.6770708 6.467442 3.8457779 3.118223  1.336736  1.2317494 12.206037
#> 6  7.2835995 6.104749 3.4486546 3.006329  1.153788  0.9860947 11.828429
#> 7  9.8164647 8.571199 6.0066878 4.874422  2.692126  3.3507985 14.312736
#> 8  6.1594154 5.232733 2.7284184 3.763176  2.152492  2.3049309 10.670921
#> 9  5.1575400 4.003579 1.3276880 2.012712  2.907885  1.4256949  9.707290
#> 10 3.4948290 2.822530 1.6451171 3.834291  4.779341  3.8272241  7.941311
#> 11 3.6003657 2.436229 0.2443488 2.280201  4.459492  2.8803877  8.142012
#> 12 1.3027844 2.625870 4.9874925 6.617260  9.027043  7.6435488  3.625690
#> 13 3.5281738 2.499464 0.5752938 2.771720  4.458152  3.0941738  8.085070
#> 14 4.1882815 3.033389 0.3546844 2.068949  3.864394  2.3178071  8.735217
#> 15 0.0000000 1.329231 3.8355978 5.336115  7.986325  6.4703164  4.557494
#> 16 1.3292314 0.000000 2.6800943 4.010828  6.893350  5.2443556  5.743108
#> 17 3.8355978 2.680094 0.0000000 2.197544  4.215159  2.6567533  8.381003
#> 18 5.3361155 4.010828 2.1975438 0.000000  4.127152  2.0333757  9.629610
#> 19 7.9863249 6.893350 4.2151594 4.127152  0.000000  2.0941071 12.542913
#> 20 6.4703164 5.244356 2.6567533 2.033376  2.094107  0.0000000 10.986276
#> 21 4.5574943 5.743108 8.3810025 9.629610 12.542913 10.9862762  0.000000
#> 22 5.1105070 4.025892 1.3696013 2.441574  2.878560  1.7137952  9.667967
#> 23 5.0330474 4.083185 1.6343384 3.130152  3.051302  2.3283550  9.565036
#> 24 2.9902546 2.521192 2.0773589 4.266323  5.423291  4.4370592  7.357544
#> 25 4.2845099 3.234676 0.7066955 2.553186  3.701833  2.4330105  8.841200
#>           22        23        24        25
#> 1  0.7824482 0.2342420 2.1540957 0.7606689
#> 2  5.7407508 5.5977165 3.3994700 4.9089343
#> 3  2.5601760 2.9726552 2.9559784 2.0624730
#> 4  0.8952548 1.6075522 3.2913648 1.2986306
#> 5  2.7256333 3.1874143 5.4793162 3.5268342
#> 6  2.2576908 2.6807292 4.9976088 3.0755125
#> 7  4.9216082 5.3561232 7.6730877 5.7211035
#> 8  1.4633966 1.1496943 3.3770861 2.0477971
#> 9  0.4405413 1.1561888 3.0147583 1.0074324
#> 10 2.1305372 1.7306128 0.6456859 1.4272679
#> 11 1.6109808 1.8388480 1.9902656 0.9089880
#> 12 6.1825565 6.0026590 3.7384478 5.3505629
#> 13 1.5844308 1.5895570 1.5033453 0.7563965
#> 14 1.0394577 1.4024954 2.2894661 0.5393182
#> 15 5.1105070 5.0330474 2.9902546 4.2845099
#> 16 4.0258915 4.0831846 2.5211917 3.2346763
#> 17 1.3696013 1.6343384 2.0773589 0.7066955
#> 18 2.4415743 3.1301516 4.2663229 2.5531861
#> 19 2.8785603 3.0513022 5.4232911 3.7018332
#> 20 1.7137952 2.3283550 4.4370592 2.4330105
#> 21 9.6679671 9.5650359 7.3575438 8.8412005
#> 22 0.0000000 0.7174232 2.7572861 0.8321875
#> 23 0.7174232 0.0000000 2.3760179 0.9304864
#> 24 2.7572861 2.3760179 0.0000000 2.0104069
#> 25 0.8321875 0.9304864 2.0104069 0.0000000