Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3         4         5         6         7
#> 1  0.0000000 3.1619098  5.627779 1.2364999 0.7061917 2.3685076 3.3500590
#> 2  3.1619098 0.0000000  8.143673 3.3402802 3.4664777 5.3945953 6.1733089
#> 3  5.6277795 8.1436732  0.000000 4.8609064 4.9905636 3.5073848 2.3780294
#> 4  1.2364999 3.3402802  4.860906 0.0000000 0.6445175 2.2020111 2.8403840
#> 5  0.7061917 3.4664777  4.990564 0.6445175 0.0000000 1.9281229 2.7842059
#> 6  2.3685076 5.3945953  3.507385 2.2020111 1.9281229 0.0000000 1.1302114
#> 7  3.3500590 6.1733089  2.378029 2.8403840 2.7842059 1.1302114 0.0000000
#> 8  1.5750154 4.7152767  4.421266 1.8452928 1.3475666 0.9240897 2.0442204
#> 9  3.0140618 1.4056717  8.522756 3.6853152 3.5565696 5.3825282 6.3350956
#> 10 2.1829206 0.9890047  7.244570 2.3982368 2.4813152 4.4088610 5.2114946
#> 11 3.2588811 6.3135643  3.016514 3.1020885 2.8472706 0.9196085 0.9333490
#> 12 0.7053608 3.8656332  5.207429 1.4965866 0.8535998 1.7929858 2.8597716
#> 13 5.0655786 2.9472027 10.619318 5.7829162 5.6425803 7.4222823 8.4094621
#> 14 1.3164322 4.4781934  4.790890 1.8323261 1.2520605 1.2950468 2.4145634
#> 15 3.0080810 0.8994653  8.375422 3.5148114 3.4741506 5.3595233 6.2558446
#> 16 2.5576206 0.6351333  7.528497 2.7089375 2.8343256 4.7622887 5.5391848
#> 17 1.5575663 4.7004250  4.957764 2.2055798 1.6070541 1.4553724 2.5841769
#> 18 0.7327503 2.5020365  6.306764 1.6388603 1.3176986 3.1004880 4.0638209
#> 19 3.9764614 6.8046060  1.873845 3.4679498 3.4194286 1.6896125 0.6358677
#> 20 4.2350555 6.5365612  1.645953 3.3191704 3.5526488 2.5109291 1.5439405
#> 21 2.2957869 5.1829178  5.819532 3.2290544 2.5928567 2.4282183 3.5055843
#> 22 3.3433009 0.6218493  8.568671 3.7185148 3.7504171 5.6629547 6.5133203
#> 23 3.9187438 2.5089028  9.531954 4.7665958 4.5420645 6.2432042 7.2669559
#> 24 1.8205107 4.7088799  3.824484 1.4615024 1.2632881 0.7447645 1.5334460
#> 25 4.0988732 6.6473566  1.535572 3.3379585 3.4555310 2.1039952 1.0203358
#>            8         9        10        11        12        13        14
#> 1  1.5750154 3.0140618 2.1829206 3.2588811 0.7053608  5.065579 1.3164322
#> 2  4.7152767 1.4056717 0.9890047 6.3135643 3.8656332  2.947203 4.4781934
#> 3  4.4212659 8.5227564 7.2445701 3.0165140 5.2074291 10.619318 4.7908904
#> 4  1.8452928 3.6853152 2.3982368 3.1020885 1.4965866  5.782916 1.8323261
#> 5  1.3475666 3.5565696 2.4813152 2.8472706 0.8535998  5.642580 1.2520605
#> 6  0.9240897 5.3825282 4.4088610 0.9196085 1.7929858  7.422282 1.2950468
#> 7  2.0442204 6.3350956 5.2114946 0.9333490 2.8597716  8.409462 2.4145634
#> 8  0.0000000 4.5558394 3.7278842 1.7238025 0.9175302  6.560184 0.3711484
#> 9  4.5558394 0.0000000 1.5384192 6.2665023 3.6395271  2.098515 4.2427695
#> 10 3.7278842 1.5384192 0.0000000 5.3272410 2.8881533  3.531065 3.4972638
#> 11 1.7238025 6.2665023 5.3272410 0.0000000 2.6384016  8.283320 2.0707368
#> 12 0.9175302 3.6395271 2.8881533 2.6384016 0.0000000  5.645098 0.6176385
#> 13 6.5601844 2.0985148 3.5310653 8.2833203 5.6450977  0.000000 6.2249044
#> 14 0.3711484 4.2427695 3.4972638 2.0707368 0.6176385  6.224904 0.0000000
#> 15 4.5793363 0.5065576 1.2002398 6.2622394 3.6758725  2.330847 4.2910957
#> 16 4.0966901 1.5129126 0.3764452 5.6815832 3.2629693  3.375890 3.8707903
#> 17 0.6137352 4.3585297 3.7318682 2.1048288 0.8607989  6.285910 0.3839857
#> 18 2.2984256 2.2830728 1.5560514 3.9913556 1.3982973  4.345660 2.0155814
#> 19 2.6135829 6.9682034 5.8452977 1.1550891 3.4635254  9.039440 2.9843097
#> 20 3.3100303 7.0044536 5.6635574 2.4663673 3.9392252  9.102029 3.6529951
#> 21 1.7240361 4.5388019 4.2838174 2.8206990 1.7442133  6.246441 1.4974258
#> 22 4.9181052 0.9262244 1.3241124 6.5746081 4.0292528  2.338128 4.6468390
#> 23 5.3597728 1.2150884 2.7506717 7.0825510 4.4528445  1.271839 5.0141331
#> 24 0.8562798 4.8031155 3.7306186 1.6416555 1.4014248  6.876044 1.1417788
#> 25 2.9776966 6.9888112 5.7310394 1.8959901 3.7085101  9.084674 3.3394678
#>           15        16        17        18        19       20       21
#> 1  3.0080810 2.5576206 1.5575663 0.7327503 3.9764614 4.235055 2.295787
#> 2  0.8994653 0.6351333 4.7004250 2.5020365 6.8046060 6.536561 5.182918
#> 3  8.3754219 7.5284973 4.9577637 6.3067637 1.8738446 1.645953 5.819532
#> 4  3.5148114 2.7089375 2.2055798 1.6388603 3.4679498 3.319170 3.229054
#> 5  3.4741506 2.8343256 1.6070541 1.3176986 3.4194286 3.552649 2.592857
#> 6  5.3595233 4.7622887 1.4553724 3.1004880 1.6896125 2.510929 2.428218
#> 7  6.2558446 5.5391848 2.5841769 4.0638209 0.6358677 1.543941 3.505584
#> 8  4.5793363 4.0966901 0.6137352 2.2984256 2.6135829 3.310030 1.724036
#> 9  0.5065576 1.5129126 4.3585297 2.2830728 6.9682034 7.004454 4.538802
#> 10 1.2002398 0.3764452 3.7318682 1.5560514 5.8452977 5.663557 4.283817
#> 11 6.2622394 5.6815832 2.1048288 3.9913556 1.1550891 2.466367 2.820699
#> 12 3.6758725 3.2629693 0.8607989 1.3982973 3.4635254 3.939225 1.744213
#> 13 2.3308471 3.3758900 6.2859101 4.3456602 9.0394396 9.102029 6.246441
#> 14 4.2910957 3.8707903 0.3839857 2.0155814 2.9843097 3.652995 1.497426
#> 15 0.0000000 1.0833881 4.4483956 2.2809758 6.8916160 6.820310 4.746593
#> 16 1.0833881 0.0000000 4.1079364 1.9309224 6.1709585 5.930767 4.653389
#> 17 4.4483956 4.1079364 0.0000000 2.2051486 3.1131251 3.911141 1.127573
#> 18 2.2809758 1.9309224 2.2051486 0.0000000 4.6940374 4.861347 2.739381
#> 19 6.8916160 6.1709585 3.1131251 4.6940374 0.0000000 1.521469 3.945712
#> 20 6.8203098 5.9307671 3.9111406 4.8613474 1.5214694 0.000000 4.938052
#> 21 4.7465932 4.6533886 1.1275729 2.7393812 3.9457118 4.938052 0.000000
#> 22 0.4727320 1.0757774 4.8267081 2.6317181 7.1486310 6.985267 5.178755
#> 23 1.6542371 2.7198687 5.0515155 3.2251902 7.8896291 8.067034 4.975722
#> 24 4.7364007 4.0740990 1.4637169 2.5304463 2.1651693 2.547195 2.580312
#> 25 6.8490615 6.0254587 3.5519443 4.7723138 0.8839129 0.639717 4.515349
#>           22       23        24        25
#> 1  3.3433009 3.918744 1.8205107 4.0988732
#> 2  0.6218493 2.508903 4.7088799 6.6473566
#> 3  8.5686708 9.531954 3.8244843 1.5355724
#> 4  3.7185148 4.766596 1.4615024 3.3379585
#> 5  3.7504171 4.542065 1.2632881 3.4555310
#> 6  5.6629547 6.243204 0.7447645 2.1039952
#> 7  6.5133203 7.266956 1.5334460 1.0203358
#> 8  4.9181052 5.359773 0.8562798 2.9776966
#> 9  0.9262244 1.215088 4.8031155 6.9888112
#> 10 1.3241124 2.750672 3.7306186 5.7310394
#> 11 6.5746081 7.082551 1.6416555 1.8959901
#> 12 4.0292528 4.452844 1.4014248 3.7085101
#> 13 2.3381282 1.271839 6.8760437 9.0846740
#> 14 4.6468390 5.014133 1.1417788 3.3394678
#> 15 0.4727320 1.654237 4.7364007 6.8490615
#> 16 1.0757774 2.719869 4.0740990 6.0254587
#> 17 4.8267081 5.051516 1.4637169 3.5519443
#> 18 2.6317181 3.225190 2.5304463 4.7723138
#> 19 7.1486310 7.889629 2.1651693 0.8839129
#> 20 6.9852670 8.067034 2.5471951 0.6397170
#> 21 5.1787548 4.975722 2.5803121 4.5153490
#> 22 0.0000000 1.915088 5.0120616 7.0540728
#> 23 1.9150882 0.000000 5.7392538 7.9974028
#> 24 5.0120616 5.739254 0.0000000 2.3115795
#> 25 7.0540728 7.997403 2.3115795 0.0000000