Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1          2         3        4         5         6         7
#> 1  0.0000000  0.6538297 4.8091804 5.343620 2.6979033 4.3008316  3.063618
#> 2  0.6538297  0.0000000 5.4619558 5.953486 3.2231609 4.8920823  2.861824
#> 3  4.8091804  5.4619558 0.0000000 2.328114 3.4250602 2.4489129  6.786839
#> 4  5.3436202  5.9534861 2.3281135 0.000000 2.9039779 1.1147804  6.230651
#> 5  2.6979033  3.2231609 3.4250602 2.903978 0.0000000 1.7899581  3.455710
#> 6  4.3008316  4.8920823 2.4489129 1.114780 1.7899581 0.0000000  5.140598
#> 7  3.0636179  2.8618240 6.7868392 6.230651 3.4557095 5.1405976  0.000000
#> 8  3.3272225  3.9801390 1.5556527 2.452105 2.0116921 1.8367284  5.256005
#> 9  3.8294194  4.4770845 1.3277704 1.865850 2.1016642 1.3907895  5.502266
#> 10 1.8797168  2.4697613 3.3907066 3.486453 0.9247130 2.4241651  3.433834
#> 11 3.7543520  4.3889925 1.3602481 3.284356 3.1496473 2.9405101  6.160244
#> 12 6.5646939  7.2077105 2.2511355 1.697135 4.4230473 2.7108527  7.842442
#> 13 9.6761012 10.3288490 4.8669264 5.389997 7.9555182 6.3556264 11.411069
#> 14 5.8035788  6.4472877 1.6351561 1.226230 3.7150501 2.0726755  7.156956
#> 15 2.3757491  2.9771983 2.9798782 2.976530 0.7153159 1.9259951  3.807192
#> 16 3.6831062  4.2694681 2.5721900 1.725578 1.1876208 0.6248359  4.601237
#> 17 2.2267379  2.8591472 2.8254946 3.135382 1.1149623 2.1480916  4.006557
#> 18 2.4690455  3.0358608 3.2010770 2.959578 0.3900315 1.8677201  3.600442
#> 19 3.8827170  4.5242276 1.1292165 3.064297 3.1002355 2.7672149  6.198727
#> 20 4.2016151  4.8536162 0.8774239 1.855406 2.5625096 1.6590416  5.962052
#> 21 9.4732491 10.1220305 4.6866054 5.511520 7.9202436 6.4058605 11.369076
#> 22 7.1592125  7.8092054 2.3702750 3.576440 5.6729243 4.2935874  9.101139
#> 23 4.9624724  5.6158943 0.1899382 2.219363 3.4945644 2.4230585  6.882396
#> 24 4.6378250  5.2883124 0.2823064 2.554525 3.4092717 2.5786301  6.718887
#> 25 2.8917485  3.1993607 4.7357312 3.888741 1.3190124 2.8158296  2.350801
#>            8         9        10        11        12         13        14
#> 1  3.3272225 3.8294194 1.8797168 3.7543520 6.5646939  9.6761012 5.8035788
#> 2  3.9801390 4.4770845 2.4697613 4.3889925 7.2077105 10.3288490 6.4472877
#> 3  1.5556527 1.3277704 3.3907066 1.3602481 2.2511355  4.8669264 1.6351561
#> 4  2.4521053 1.8658503 3.4864530 3.2843559 1.6971352  5.3899974 1.2262299
#> 5  2.0116921 2.1016642 0.9247130 3.1496473 4.4230473  7.9555182 3.7150501
#> 6  1.8367284 1.3907895 2.4241651 2.9405101 2.7108527  6.3556264 2.0726755
#> 7  5.2560052 5.5022664 3.4338343 6.1602440 7.8424415 11.4110692 7.1569557
#> 8  0.0000000 0.5977412 1.8388376 1.2071070 3.2966139  6.3836064 2.5393209
#> 9  0.5977412 0.0000000 2.1812641 1.5517299 2.7434680  5.9818660 1.9819004
#> 10 1.8388376 2.1812641 0.0000000 2.7526699 4.8203412  8.1625583 4.0706494
#> 11 1.2071070 1.5517299 2.7526699 0.0000000 3.5902273  6.0470756 2.9179863
#> 12 3.2966139 2.7434680 4.8203412 3.5902273 0.0000000  3.6935027 0.7617754
#> 13 6.3836064 5.9818660 8.1625583 6.0470756 3.6935027  0.0000000 4.2844787
#> 14 2.5393209 1.9819004 4.0706494 2.9179863 0.7617754  4.2844787 0.0000000
#> 15 1.4593069 1.7188600 0.5126673 2.5084253 4.3150897  7.6955994 3.5686929
#> 16 1.5631263 1.3065212 1.8040926 2.7564358 3.2462413  6.8290205 2.5570289
#> 17 1.2705341 1.6560710 0.5734750 2.1999955 4.3549943  7.6237442 3.5970015
#> 18 1.7212297 1.9020777 0.6078777 2.8142686 4.3877193  7.8427699 3.6560075
#> 19 1.1091222 1.3769225 2.7724219 0.2393711 3.3521681  5.8663543 2.6786271
#> 20 0.8784643 0.4620164 2.6193961 1.4290242 2.4332451  5.5465223 1.6842848
#> 21 6.2342043 5.8841310 8.0492995 5.7747432 3.8410301  0.6919566 4.3421816
#> 22 3.9216407 3.5995945 5.7459374 3.4921241 2.1604624  2.5563641 2.3505097
#> 23 1.6805264 1.3932166 3.5046212 1.5499358 2.0641399  4.7160644 1.4623185
#> 24 1.4632827 1.3481424 3.2993629 1.0944316 2.5325761  5.0487808 1.9147197
#> 25 3.3253601 3.4086226 1.8242594 4.4333340 5.5266259  9.1485656 4.8681457
#>           15        16        17        18        19        20         21
#> 1  2.3757491 3.6831062 2.2267379 2.4690455 3.8827170 4.2016151  9.4732491
#> 2  2.9771983 4.2694681 2.8591472 3.0358608 4.5242276 4.8536162 10.1220305
#> 3  2.9798782 2.5721900 2.8254946 3.2010770 1.1292165 0.8774239  4.6866054
#> 4  2.9765304 1.7255782 3.1353818 2.9595777 3.0642971 1.8554061  5.5115204
#> 5  0.7153159 1.1876208 1.1149623 0.3900315 3.1002355 2.5625096  7.9202436
#> 6  1.9259951 0.6248359 2.1480916 1.8677201 2.7672149 1.6590416  6.4058605
#> 7  3.8071915 4.6012370 4.0065574 3.6004423 6.1987266 5.9620516 11.3690759
#> 8  1.4593069 1.5631263 1.2705341 1.7212297 1.1091222 0.8784643  6.2342043
#> 9  1.7188600 1.3065212 1.6560710 1.9020777 1.3769225 0.4620164  5.8841310
#> 10 0.5126673 1.8040926 0.5734750 0.6078777 2.7724219 2.6193961  8.0492995
#> 11 2.5084253 2.7564358 2.1999955 2.8142686 0.2393711 1.4290242  5.7747432
#> 12 4.3150897 3.2462413 4.3549943 4.3877193 3.3521681 2.4332451  3.8410301
#> 13 7.6955994 6.8290205 7.6237442 7.8427699 5.8663543 5.5465223  0.6919566
#> 14 3.5686929 2.5570289 3.5970015 3.6560075 2.6786271 1.6842848  4.3421816
#> 15 0.0000000 1.3132923 0.4016006 0.3293168 2.4881635 2.1703594  7.6024739
#> 16 1.3132923 0.0000000 1.5693136 1.2429187 2.6229779 1.7043405  6.8340094
#> 17 0.4016006 1.5693136 0.0000000 0.7257725 2.2075331 2.0774949  7.4954524
#> 18 0.3293168 1.2429187 0.7257725 0.0000000 2.7817918 2.3626175  7.7752348
#> 19 2.4881635 2.6229779 2.2075331 2.7817918 0.0000000 1.2118674  5.6123376
#> 20 2.1703594 1.7043405 2.0774949 2.3626175 1.2118674 0.0000000  5.4334070
#> 21 7.6024739 6.8340094 7.4954524 7.7752348 5.6123376 5.4334070  0.0000000
#> 22 5.3110936 4.6448120 5.1871661 5.5007272 3.3160188 3.1419193  2.3174813
#> 23 3.0793775 2.5923331 2.9446604 3.2873818 1.3182526 0.9321042  4.5536780
#> 24 2.9165583 2.6379902 2.7278447 3.1589251 0.8703410 0.9382340  4.8373357
#> 25 1.9278061 2.3225404 2.2764024 1.6198506 4.4005733 3.8663592  9.1549819
#>          22        23        24       25
#> 1  7.159212 4.9624724 4.6378250 2.891749
#> 2  7.809205 5.6158943 5.2883124 3.199361
#> 3  2.370275 0.1899382 0.2823064 4.735731
#> 4  3.576440 2.2193633 2.5545248 3.888741
#> 5  5.672924 3.4945644 3.4092717 1.319012
#> 6  4.293587 2.4230585 2.5786301 2.815830
#> 7  9.101139 6.8823958 6.7188868 2.350801
#> 8  3.921641 1.6805264 1.4632827 3.325360
#> 9  3.599594 1.3932166 1.3481424 3.408623
#> 10 5.745937 3.5046212 3.2993629 1.824259
#> 11 3.492124 1.5499358 1.0944316 4.433334
#> 12 2.160462 2.0641399 2.5325761 5.526626
#> 13 2.556364 4.7160644 5.0487808 9.148566
#> 14 2.350510 1.4623185 1.9147197 4.868146
#> 15 5.311094 3.0793775 2.9165583 1.927806
#> 16 4.644812 2.5923331 2.6379902 2.322540
#> 17 5.187166 2.9446604 2.7278447 2.276402
#> 18 5.500727 3.2873818 3.1589251 1.619851
#> 19 3.316019 1.3182526 0.8703410 4.400573
#> 20 3.141919 0.9321042 0.9382340 3.866359
#> 21 2.317481 4.5536780 4.8373357 9.154982
#> 22 0.000000 2.2425127 2.5215639 6.942957
#> 23 2.242513 0.0000000 0.4686505 4.797474
#> 24 2.521564 0.4686505 0.0000000 4.727101
#> 25 6.942957 4.7974738 4.7271007 0.000000