Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1        2         3         4         5         6         7
#> 1   0.000000 3.587210 3.0693593 3.5939934 4.3857009  9.344569 1.5192537
#> 2   3.587210 0.000000 1.6336550 2.0537686 3.0767289  7.500723 2.3281223
#> 3   3.069359 1.633655 0.0000000 0.6369202 1.7052607  6.601333 1.5502298
#> 4   3.593993 2.053769 0.6369202 0.0000000 1.0866702  5.967856 2.0929216
#> 5   4.385701 3.076729 1.7052607 1.0866702 0.0000000  4.991079 2.9722199
#> 6   9.344569 7.500723 6.6013329 5.9678556 4.9910787  0.000000 7.9631700
#> 7   1.519254 2.328122 1.5502298 2.0929216 2.9722199  7.963170 0.0000000
#> 8   2.935568 2.840414 3.7571366 4.3823142 5.4622298 10.246065 2.9841544
#> 9   1.746315 2.836394 1.6378122 2.0042814 2.6614003  7.599573 0.7298774
#> 10  3.987504 2.264444 1.0196996 0.3994320 0.8128570  5.582740 2.4908015
#> 11  6.208579 3.627673 3.1526229 2.6221907 2.1810815  3.958436 4.6950698
#> 12  6.752000 5.257724 4.1228849 3.4869821 2.4455167  2.600197 5.3997783
#> 13  3.958039 1.167847 1.0800492 1.1318682 1.9981983  6.353453 2.4729502
#> 14  2.730300 1.350421 2.2967395 2.8993560 3.9859832  8.710668 1.9916251
#> 15  2.508860 1.924505 2.7408461 3.3645002 4.4454315  9.239781 2.1174857
#> 16  2.479418 1.889667 2.6874890 3.3117722 4.3922548  9.190938 2.0658148
#> 17  6.262541 3.583984 3.1973744 2.6899834 2.3146665  4.062236 4.7454479
#> 18  5.830594 3.762490 2.8897693 2.2767415 1.5221999  3.764527 4.3566217
#> 19  7.393678 5.820953 4.7437381 4.1069023 3.0792619  1.964029 6.0393729
#> 20  4.932991 3.370859 2.1589220 1.5223281 0.5520405  4.467676 3.5042683
#> 21  4.626394 2.411028 1.5787789 1.0448742 1.0493655  5.151144 3.1140794
#> 22 10.433705 8.970608 7.9097725 7.2736795 6.2263911  1.688193 9.1506390
#> 23  8.045957 6.722219 5.5578043 4.9254993 3.8599459  1.653965 6.7628690
#> 24  1.663510 2.280841 2.5467552 3.1817582 4.2126867  9.147953 1.4882705
#> 25  2.054137 1.656289 2.0045694 2.6404856 3.7024557  8.585410 1.3147055
#>            8         9        10        11        12       13         14
#> 1   2.935568 1.7463152 3.9875038 6.2085785 6.7520002 3.958039  2.7302998
#> 2   2.840414 2.8363941 2.2644440 3.6276730 5.2577245 1.167847  1.3504206
#> 3   3.757137 1.6378122 1.0196996 3.1526229 4.1228849 1.080049  2.2967395
#> 4   4.382314 2.0042814 0.3994320 2.6221907 3.4869821 1.131868  2.8993560
#> 5   5.462230 2.6614003 0.8128570 2.1810815 2.4455167 1.998198  3.9859832
#> 6  10.246065 7.5995733 5.5827402 3.9584365 2.6001971 6.353453  8.7106676
#> 7   2.984154 0.7298774 2.4908015 4.6950698 5.3997783 2.472950  1.9916251
#> 8   0.000000 3.7027793 4.7280627 6.4508023 7.8561404 3.914750  1.5377177
#> 9   3.702779 0.0000000 2.3700571 4.6100774 5.0058027 2.697817  2.6990310
#> 10  4.728063 2.3700571 0.0000000 2.2412769 3.1284240 1.200371  3.2239986
#> 11  6.450802 4.6100774 2.2412769 0.0000000 2.2790324 2.550940  4.9208578
#> 12  7.856140 5.0058027 3.1284240 2.2790324 0.0000000 4.090905  6.3443019
#> 13  3.914750 2.6978175 1.2003707 2.5509400 4.0909047 0.000000  2.3784143
#> 14  1.537718 2.6990310 3.2239986 4.9208578 6.3443019 2.378414  0.0000000
#> 15  1.018107 2.8472858 3.7102969 5.4806615 6.8385799 2.932189  0.5741196
#> 16  1.070559 2.7955099 3.6585585 5.4370397 6.7869317 2.887724  0.5400484
#> 17  6.417085 4.6881568 2.3186060 0.1899026 2.4589450 2.537334  4.8931413
#> 18  6.481539 4.1430019 1.8781372 0.8967803 1.5969857 2.601368  4.9464321
#> 19  8.461958 5.6474313 3.7399561 2.6149171 0.6418704 4.653361  6.9423088
#> 20  5.893613 3.2126626 1.1663057 1.7408527 1.9653906 2.227948  4.3895564
#> 21  5.094999 3.0460918 0.6850744 1.5822547 2.8478754 1.243403  3.5605605
#> 22 11.634726 8.6969758 6.9112424 5.5411496 3.7869293 7.806198 10.1139875
#> 23  9.304926 6.3073426 4.5783411 3.5708470 1.4646835 5.555144  7.8003097
#> 24  1.513216 2.1955067 3.5653547 5.5627531 6.6544938 3.045946  1.1278514
#> 25  1.796158 2.0399810 3.0099176 4.9438610 6.1273493 2.413775  0.7192721
#>             15          16        17        18        19        20        21
#> 1   2.50886027  2.47941799 6.2625413 5.8305942 7.3936775 4.9329910 4.6263936
#> 2   1.92450475  1.88966676 3.5839843 3.7624896 5.8209527 3.3708588 2.4110284
#> 3   2.74084612  2.68748898 3.1973744 2.8897693 4.7437381 2.1589220 1.5787789
#> 4   3.36450019  3.31177219 2.6899834 2.2767415 4.1069023 1.5223281 1.0448742
#> 5   4.44543152  4.39225483 2.3146665 1.5221999 3.0792619 0.5520405 1.0493655
#> 6   9.23978116  9.19093817 4.0622363 3.7645267 1.9640287 4.4676762 5.1511442
#> 7   2.11748567  2.06581483 4.7454479 4.3566217 6.0393729 3.5042683 3.1140794
#> 8   1.01810683  1.07055902 6.4170852 6.4815390 8.4619578 5.8936133 5.0949993
#> 9   2.84728584  2.79550995 4.6881568 4.1430019 5.6474313 3.2126626 3.0460918
#> 10  3.71029693  3.65855846 2.3186060 1.8781372 3.7399561 1.1663057 0.6850744
#> 11  5.48066154  5.43703967 0.1899026 0.8967803 2.6149171 1.7408527 1.5822547
#> 12  6.83857994  6.78693170 2.4589450 1.5969857 0.6418704 1.9653906 2.8478754
#> 13  2.93218889  2.88772382 2.5373342 2.6013684 4.6533606 2.2279477 1.2434028
#> 14  0.57411957  0.54004838 4.8931413 4.9464321 6.9423088 4.3895564 3.5605605
#> 15  0.00000000  0.05508537 5.4572030 5.4760141 7.4457028 4.8756445 4.0894764
#> 16  0.05508537  0.00000000 5.4147712 5.4274708 7.3946354 4.8236868 4.0410727
#> 17  5.45720295  5.41477117 0.0000000 1.0826138 2.7773201 1.8955279 1.6451735
#> 18  5.47601412  5.42747076 1.0826138 0.0000000 2.0847585 0.9930027 1.3867010
#> 19  7.44570277  7.39463538 2.7773201 2.0847585 0.0000000 2.5850888 3.4121154
#> 20  4.87564452  4.82368676 1.8955279 0.9930027 2.5850888 0.0000000 1.0550026
#> 21  4.08947641  4.04107274 1.6451735 1.3867010 3.4121154 1.0550026 0.0000000
#> 22 10.61842112 10.56730350 5.6630378 5.2091262 3.1727691 5.7515690 6.5716254
#> 23  8.28683600  8.23464709 3.7309998 3.0228268 0.9575185 3.4120515 4.3118738
#> 24  0.84768077  0.81627655 5.5682502 5.4199938 7.2818908 4.6981345 4.0596525
#> 25  0.82864817  0.77425338 4.9437656 4.8384885 6.7465942 4.1627574 3.4648587
#>           22        23         24        25
#> 1  10.433705 8.0459575  1.6635105 2.0541375
#> 2   8.970608 6.7222192  2.2808409 1.6562890
#> 3   7.909773 5.5578043  2.5467552 2.0045694
#> 4   7.273680 4.9254993  3.1817582 2.6404856
#> 5   6.226391 3.8599459  4.2126867 3.7024557
#> 6   1.688193 1.6539655  9.1479531 8.5854100
#> 7   9.150639 6.7628690  1.4882705 1.3147055
#> 8  11.634726 9.3049262  1.5132165 1.7961583
#> 9   8.696976 6.3073426  2.1955067 2.0399810
#> 10  6.911242 4.5783411  3.5653547 3.0099176
#> 11  5.541150 3.5708470  5.5627531 4.9438610
#> 12  3.786929 1.4646835  6.6544938 6.1273493
#> 13  7.806198 5.5551437  3.0459460 2.4137753
#> 14 10.113987 7.8003097  1.1278514 0.7192721
#> 15 10.618421 8.2868360  0.8476808 0.8286482
#> 16 10.567304 8.2346471  0.8162765 0.7742534
#> 17  5.663038 3.7309998  5.5682502 4.9437656
#> 18  5.209126 3.0228268  5.4199938 4.8384885
#> 19  3.172769 0.9575185  7.2818908 6.7465942
#> 20  5.751569 3.4120515  4.6981345 4.1627574
#> 21  6.571625 4.3118738  4.0596525 3.4648587
#> 22  0.000000 2.3905042 10.4388178 9.9141478
#> 23  2.390504 0.0000000  8.0696264 7.5606120
#> 24 10.438818 8.0696264  0.0000000 0.6406948
#> 25  9.914148 7.5606120  0.6406948 0.0000000