Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3        4         5         6         7
#> 1  0.0000000  7.178445  7.295838 5.641943 4.1865678  8.713385 3.6418365
#> 2  7.1784447  0.000000  3.402994 1.550869 3.1930369  3.116875 4.9475990
#> 3  7.2958379  3.402994  0.000000 3.259134 3.4481830  1.754131 3.8191418
#> 4  5.6419427  1.550869  3.259134 0.000000 1.6747830  3.764145 3.5441040
#> 5  4.1865678  3.193037  3.448183 1.674783 0.0000000  4.605969 1.9582552
#> 6  8.7133847  3.116875  1.754131 3.764145 4.6059694  0.000000 5.4289453
#> 7  3.6418365  4.947599  3.819142 3.544104 1.9582552  5.428945 0.0000000
#> 8  3.3929943  3.999693  4.042281 2.461757 0.8237448  5.339151 1.5132471
#> 9  3.7057545  4.424174  3.605250 2.994166 1.3953394  5.126220 0.5629359
#> 10 2.3367755  4.954191  5.039289 3.403322 1.8572674  6.388858 1.8099493
#> 11 3.2283775  4.259616  4.125095 2.729073 1.0687586  5.485378 1.2879508
#> 12 2.0519142  5.610708  6.577517 4.173647 3.1391656  7.698861 3.6581575
#> 13 4.0115189  3.561035  3.410013 2.077664 0.4484960  4.707834 1.5110176
#> 14 2.1782039  5.034176  5.279306 3.486939 2.0154974  6.591113 2.1026330
#> 15 0.2852376  7.341387  7.546786 5.813418 4.3948511  8.944741 3.9161144
#> 16 2.1046708  7.587974  6.716637 6.061942 4.3953361  8.356100 2.9287654
#> 17 3.6258755  6.672830  5.086668 5.282193 3.6741632  6.808972 1.7382293
#> 18 1.3717763  8.481433  8.291648 6.934656 5.4035539  9.809238 4.4976303
#> 19 5.3643586  2.778599  2.104239 1.708904 1.3446251  3.349132 2.3085294
#> 20 1.3075310  6.315537  6.061375 4.765328 3.1705914  7.547547 2.3423737
#> 21 5.4159465  2.381183  2.320647 1.258316 1.2556642  3.351382 2.6048230
#> 22 3.1510178 10.273619 10.379110 8.755373 7.3371428 11.848110 6.6115895
#> 23 4.4158127  3.811791  5.915831 2.774318 2.8552103  6.532354 4.4667280
#> 24 4.7493873  3.120688  2.671357 1.804040 0.8182988  3.964297 1.8335586
#> 25 1.1977062  6.032992  6.466993 4.513479 3.1690522  7.766516 3.1151228
#>            8         9        10        11        12        13        14
#> 1  3.3929943 3.7057545 2.3367755 3.2283775 2.0519142 4.0115189 2.1782039
#> 2  3.9996934 4.4241741 4.9541911 4.2596160 5.6107077 3.5610351 5.0341764
#> 3  4.0422812 3.6052503 5.0392892 4.1250950 6.5775168 3.4100129 5.2793065
#> 4  2.4617574 2.9941665 3.4033221 2.7290734 4.1736475 2.0776638 3.4869388
#> 5  0.8237448 1.3953394 1.8572674 1.0687586 3.1391656 0.4484960 2.0154974
#> 6  5.3391511 5.1262196 6.3888577 5.4853781 7.6988614 4.7078337 6.5911132
#> 7  1.5132471 0.5629359 1.8099493 1.2879508 3.6581575 1.5110176 2.1026330
#> 8  0.0000000 1.0227720 1.0563279 0.2848157 2.5879793 0.6378004 1.2541499
#> 9  1.0227720 0.0000000 1.5977498 0.8544651 3.4070213 0.9481505 1.8833501
#> 10 1.0563279 1.5977498 0.0000000 0.9154705 1.8482248 1.6810645 0.2931469
#> 11 0.2848157 0.8544651 0.9154705 0.0000000 2.6028377 0.7896229 1.1587819
#> 12 2.5879793 3.4070213 1.8482248 2.6028377 0.0000000 3.1891132 1.5555967
#> 13 0.6378004 0.9481505 1.6810645 0.7896229 3.1891132 0.0000000 1.8909645
#> 14 1.2541499 1.8833501 0.2931469 1.1587819 1.5555967 1.8909645 0.0000000
#> 15 3.6129468 3.9655654 2.5578096 3.4603538 2.0684274 4.2380640 2.3797842
#> 16 3.6028393 3.3013202 2.8111196 3.3328800 3.7300730 4.0406606 2.8831003
#> 17 3.0943993 2.2915213 2.8794662 2.8203596 4.5162969 3.2329684 3.1145297
#> 18 4.5857357 4.6964074 3.5524801 4.3812106 3.3966793 5.1698217 3.4477300
#> 19 1.9995063 1.8763908 3.0420661 2.1362494 4.4804279 1.3632166 3.2536055
#> 20 2.3470135 2.4561262 1.3689887 2.1255569 2.2171809 2.9106152 1.3627696
#> 21 2.0230629 2.1196118 3.0791713 2.2100303 4.3650174 1.4214962 3.2596069
#> 22 6.5429284 6.7741026 5.4871473 6.3703263 4.7826547 7.1570850 5.3252736
#> 23 2.9575737 3.9737189 3.0906209 3.1826037 2.3972022 3.2080789 2.9275938
#> 24 1.3891233 1.3403019 2.4273103 1.5211097 3.9085542 0.7513875 2.6421713
#> 25 2.4389922 3.0066501 1.4352780 2.3432131 0.9991612 3.0767274 1.1883583
#>           15       16       17       18        19       20        21        22
#> 1  0.2852376 2.104671 3.625876 1.371776 5.3643586 1.307531 5.4159465  3.151018
#> 2  7.3413874 7.587974 6.672830 8.481433 2.7785987 6.315537 2.3811831 10.273619
#> 3  7.5467861 6.716637 5.086668 8.291648 2.1042395 6.061375 2.3206473 10.379110
#> 4  5.8134183 6.061942 5.282193 6.934656 1.7089039 4.765328 1.2583157  8.755373
#> 5  4.3948511 4.395336 3.674163 5.403554 1.3446251 3.170591 1.2556642  7.337143
#> 6  8.9447413 8.356100 6.808972 9.809238 3.3491318 7.547547 3.3513824 11.848110
#> 7  3.9161144 2.928765 1.738229 4.497630 2.3085294 2.342374 2.6048230  6.611589
#> 8  3.6129468 3.602839 3.094399 4.585736 1.9995063 2.347014 2.0230629  6.542928
#> 9  3.9655654 3.301320 2.291521 4.696407 1.8763908 2.456126 2.1196118  6.774103
#> 10 2.5578096 2.811120 2.879466 3.552480 3.0420661 1.368989 3.0791713  5.487147
#> 11 3.4603538 3.332880 2.820360 4.381211 2.1362494 2.125557 2.2100303  6.370326
#> 12 2.0684274 3.730073 4.516297 3.396679 4.4804279 2.217181 4.3650174  4.782655
#> 13 4.2380640 4.040661 3.232968 5.169822 1.3632166 2.910615 1.4214962  7.157085
#> 14 2.3797842 2.883100 3.114530 3.447730 3.2536055 1.362770 3.2596069  5.325274
#> 15 0.0000000 2.338841 3.906132 1.331614 5.5960387 1.588112 5.6332621  2.949347
#> 16 2.3388410 0.000000 1.959682 2.015347 5.1776948 1.576759 5.3986084  4.078554
#> 17 3.9061318 1.959682 0.000000 3.941318 3.9676371 2.470034 4.3076177  6.034693
#> 18 1.3316145 2.015347 3.941318 0.000000 6.4832458 2.266018 6.5911188  2.131893
#> 19 5.5960387 5.177695 3.967637 6.483246 0.0000000 4.217270 0.4518969  8.502301
#> 20 1.5881120 1.576759 2.470034 2.266018 4.2172699 0.000000 4.3317794  4.318445
#> 21 5.6332621 5.398608 4.307618 6.591119 0.4518969 4.331779 0.0000000  8.565977
#> 22 2.9493469 4.078554 6.034693 2.131893 8.5023006 4.318445 8.5659769  0.000000
#> 23 4.4592421 5.765009 5.915472 5.779528 3.9580168 4.191630 3.6242483  7.125279
#> 24 4.9807878 4.625248 3.552165 5.877009 0.6152522 3.611261 0.7793106  7.888714
#> 25 1.3157863 2.733506 3.672369 2.567655 4.4375857 1.259688 4.4231049  4.242407
#>          23        24        25
#> 1  4.415813 4.7493873 1.1977062
#> 2  3.811791 3.1206881 6.0329919
#> 3  5.915831 2.6713574 6.4669932
#> 4  2.774318 1.8040404 4.5134788
#> 5  2.855210 0.8182988 3.1690522
#> 6  6.532354 3.9642969 7.7665159
#> 7  4.466728 1.8335586 3.1151228
#> 8  2.957574 1.3891233 2.4389922
#> 9  3.973719 1.3403019 3.0066501
#> 10 3.090621 2.4273103 1.4352780
#> 11 3.182604 1.5211097 2.3432131
#> 12 2.397202 3.9085542 0.9991612
#> 13 3.208079 0.7513875 3.0767274
#> 14 2.927594 2.6421713 1.1883583
#> 15 4.459242 4.9807878 1.3157863
#> 16 5.765009 4.6252479 2.7335057
#> 17 5.915472 3.5521647 3.6723693
#> 18 5.779528 5.8770094 2.5676546
#> 19 3.958017 0.6152522 4.4375857
#> 20 4.191630 3.6112607 1.2596879
#> 21 3.624248 0.7793106 4.4231049
#> 22 7.125279 7.8887145 4.2424073
#> 23 0.000000 3.6156280 3.2506271
#> 24 3.615628 0.0000000 3.8281017
#> 25 3.250627 3.8281017 0.0000000