Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>           1        2          3         4         5         6         7
#> 1  0.000000 7.154261 1.62699123 1.6106062 5.8074773 6.2446609 4.1285395
#> 2  7.154261 0.000000 7.20944029 8.2248040 2.8432056 1.4714414 3.3195704
#> 3  1.626991 7.209440 0.00000000 1.1490492 6.4931089 6.5985081 3.9137647
#> 4  1.610606 8.224804 1.14904918 0.0000000 7.2403932 7.5022959 4.9742260
#> 5  5.807477 2.843206 6.49310887 7.2403932 0.0000000 1.3795049 3.5201253
#> 6  6.244661 1.471441 6.59850810 7.5022959 1.3795049 0.0000000 3.0114431
#> 7  4.128540 3.319570 3.91376472 4.9742260 3.5201253 3.0114431 0.0000000
#> 8  8.059553 2.504800 8.57659394 9.4200497 2.3211305 2.0535117 5.0638744
#> 9  4.756526 2.481280 5.07268092 5.9767865 1.8106150 1.5281538 1.7122089
#> 10 5.827561 1.756232 6.20292894 7.0946490 1.2340552 0.4182766 2.6971437
#> 11 2.734858 9.846783 3.55579624 2.5160954 8.2133759 8.8426062 6.8617399
#> 12 6.040912 4.920289 7.16184802 7.6312293 2.1061861 3.4771342 5.1486802
#> 13 4.481904 2.695785 4.74184280 5.6634998 2.1301865 1.8572700 1.3930784
#> 14 4.913675 2.303395 5.19205207 6.1125676 1.8335136 1.4083391 1.7224595
#> 15 1.660432 7.169002 0.05637978 1.2038236 6.4714836 6.5664566 3.8705704
#> 16 6.807828 1.533923 7.20904415 8.0971954 1.4927008 0.6241355 3.6309229
#> 17 1.284811 7.783249 0.81151286 0.4448508 6.8078966 7.0579216 4.5415401
#> 18 8.103774 1.358316 8.34152388 9.3006460 2.9183804 1.8872057 4.5348741
#> 19 6.418920 1.418171 6.77666261 7.6801446 1.4280840 0.1782493 3.1772829
#> 20 2.892498 4.589114 2.62062739 3.6661436 4.2323084 4.0691698 1.3086662
#> 21 3.769387 3.881737 4.51020183 5.2066289 2.0409239 2.6488905 2.4330889
#> 22 5.188052 3.503116 6.01238538 6.6781691 0.7906366 2.0373129 3.4796426
#> 23 4.811042 2.727131 4.57127824 5.6463437 3.4202753 2.6597802 0.6851649
#> 24 2.926255 4.697920 3.82992328 4.4210657 2.8919850 3.5237357 2.6794874
#> 25 3.124855 4.852283 4.15784781 4.6723709 2.8003754 3.5937432 3.0708398
#>            8         9        10        11       12        13        14
#> 1   8.059553 4.7565257 5.8275605  2.734858 6.040912 4.4819038 4.9136753
#> 2   2.504800 2.4812798 1.7562323  9.846783 4.920289 2.6957853 2.3033950
#> 3   8.576594 5.0726809 6.2029289  3.555796 7.161848 4.7418428 5.1920521
#> 4   9.420050 5.9767865 7.0946490  2.516095 7.631229 5.6634998 6.1125676
#> 5   2.321130 1.8106150 1.2340552  8.213376 2.106186 2.1301865 1.8335136
#> 6   2.053512 1.5281538 0.4182766  8.842606 3.477134 1.8572700 1.4083391
#> 7   5.063874 1.7122089 2.6971437  6.861740 5.148680 1.3930784 1.7224595
#> 8   0.000000 3.5252691 2.4038545 10.525388 3.639139 3.8700425 3.4325520
#> 9   3.525269 0.0000000 1.1308680  7.404623 3.593108 0.3496059 0.1829228
#> 10  2.403855 1.1308680 0.0000000  8.424728 3.330486 1.4699473 1.0287897
#> 11 10.525388 7.4046233 8.4247277  0.000000 7.943861 7.1551209 7.5711415
#> 12  3.639139 3.5931080 3.3304865  7.943861 0.000000 3.8206165 3.6872659
#> 13  3.870042 0.3496059 1.4699473  7.155121 3.820616 0.0000000 0.4506393
#> 14  3.432552 0.1829228 1.0287897  7.571141 3.687266 0.4506393 0.0000000
#> 15  8.547993 5.0413606 6.1718832  3.611777 7.156230 4.7094621 5.1594684
#> 16  1.433179 2.1365155 1.0062493  9.370329 3.455861 2.4723437 2.0260188
#> 17  8.979276 5.5321738 6.6505890  2.746421 7.245845 5.2186545 5.6677469
#> 18  1.462570 3.3491450 2.3028270 10.724585 4.732542 3.6429943 3.1988703
#> 19  1.886662 1.7063843 0.5913759  9.011421 3.504304 2.0352762 1.5860959
#> 20  6.100955 2.5805972 3.6975970  5.607204 5.451237 2.2327489 2.6688799
#> 21  4.302547 1.4716094 2.2411398  6.244183 2.894639 1.4389341 1.6522736
#> 22  3.078196 1.8825473 1.7737761  7.505059 1.710933 2.1213056 1.9775827
#> 23  4.677065 1.6566551 2.4154721  7.545043 5.242462 1.4308747 1.5933832
#> 24  5.179748 2.2274279 3.1113357  5.357431 3.350549 2.0922911 2.4096530
#> 25  5.117658 2.4270286 3.1936659  5.414079 3.004106 2.3463595 2.6094604
#>            15        16        17        18        19       20        21
#> 1  1.66043190 6.8078278 1.2848105  8.103774 6.4189196 2.892498 3.7693869
#> 2  7.16900183 1.5339229 7.7832489  1.358316 1.4181715 4.589114 3.8817374
#> 3  0.05637978 7.2090441 0.8115129  8.341524 6.7766626 2.620627 4.5102018
#> 4  1.20382362 8.0971954 0.4448508  9.300646 7.6801446 3.666144 5.2066289
#> 5  6.47148362 1.4927008 6.8078966  2.918380 1.4280840 4.232308 2.0409239
#> 6  6.56645663 0.6241355 7.0579216  1.887206 0.1782493 4.069170 2.6488905
#> 7  3.87057037 3.6309229 4.5415401  4.534874 3.1772829 1.308666 2.4330889
#> 8  8.54799305 1.4331791 8.9792760  1.462570 1.8866621 6.100955 4.3025470
#> 9  5.04136061 2.1365155 5.5321738  3.349145 1.7063843 2.580597 1.4716094
#> 10 6.17188317 1.0062493 6.6505890  2.302827 0.5913759 3.697597 2.2411398
#> 11 3.61177656 9.3703292 2.7464211 10.724585 9.0114208 5.607204 6.2441831
#> 12 7.15622977 3.4558613 7.2458453  4.732542 3.5043044 5.451237 2.8946392
#> 13 4.70946209 2.4723437 5.2186545  3.642994 2.0352762 2.232749 1.4389341
#> 14 5.15946843 2.0260188 5.6677469  3.198870 1.5860959 2.668880 1.6522736
#> 15 0.00000000 7.1778576 0.8678092  8.305067 6.7445664 2.579901 4.4934124
#> 16 7.17785763 0.0000000 7.6534684  1.447615 0.4536852 4.691290 3.1361091
#> 17 0.86780923 7.6534684 0.0000000  8.855926 7.2358035 3.232888 4.7776951
#> 18 8.30506661 1.4476151 8.8559256  0.000000 1.7253512 5.743496 4.5324014
#> 19 6.74456645 0.4536852 7.2358035  1.725351 0.0000000 4.245086 2.8071080
#> 20 2.57990121 4.6912902 3.2328879  5.743496 4.2450860 0.000000 2.5566267
#> 21 4.49341235 3.1361091 4.7776951  4.532401 2.8071080 2.556627 0.0000000
#> 22 5.99611937 2.2599794 6.2558341  3.698880 2.1250215 3.953535 1.5033136
#> 23 4.52644795 3.2544369 5.2163512  4.000930 2.8092501 1.986326 2.7478526
#> 24 3.82072737 4.0223689 4.0048972  5.410272 3.6854212 2.340486 0.8874957
#> 25 4.15264838 4.0462468 4.2702838  5.466284 3.7454122 2.776362 0.9706841
#>           22        23        24        25
#> 1  5.1880524 4.8110416 2.9262548 3.1248552
#> 2  3.5031161 2.7271313 4.6979196 4.8522835
#> 3  6.0123854 4.5712782 3.8299233 4.1578478
#> 4  6.6781691 5.6463437 4.4210657 4.6723709
#> 5  0.7906366 3.4202753 2.8919850 2.8003754
#> 6  2.0373129 2.6597802 3.5237357 3.5937432
#> 7  3.4796426 0.6851649 2.6794874 3.0708398
#> 8  3.0781961 4.6770645 5.1797476 5.1176578
#> 9  1.8825473 1.6566551 2.2274279 2.4270286
#> 10 1.7737761 2.4154721 3.1113357 3.1936659
#> 11 7.5050592 7.5450428 5.3574314 5.4140789
#> 12 1.7109331 5.2424624 3.3505486 3.0041063
#> 13 2.1213056 1.4308747 2.0922911 2.3463595
#> 14 1.9775827 1.5933832 2.4096530 2.6094604
#> 15 5.9961194 4.5264480 3.8207274 4.1526484
#> 16 2.2599794 3.2544369 4.0223689 4.0462468
#> 17 6.2558341 5.2163512 4.0048972 4.2702838
#> 18 3.6988802 4.0009296 5.4102717 5.4662844
#> 19 2.1250215 2.8092501 3.6854212 3.7454122
#> 20 3.9535350 1.9863262 2.3404864 2.7763618
#> 21 1.5033136 2.7478526 0.8874957 0.9706841
#> 22 0.0000000 3.5348169 2.2646313 2.1015925
#> 23 3.5348169 0.0000000 3.1606806 3.5155677
#> 24 2.2646313 3.1606806 0.0000000 0.4359444
#> 25 2.1015925 3.5155677 0.4359444 0.0000000