Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3          4         5         6        7
#> 1   0.000000 5.8644837 5.5453572  2.4159245 5.7322223  1.182521 7.784517
#> 2   5.864484 0.0000000 0.5102855  3.9647433 1.5085711  5.326138 1.937209
#> 3   5.545357 0.5102855 0.0000000  3.7955977 1.0862576  4.937055 2.348671
#> 4   2.415924 3.9647433 3.7955977  0.0000000 4.3720829  2.648084 5.762456
#> 5   5.732222 1.5085711 1.0862576  4.3720829 0.0000000  4.929691 2.854077
#> 6   1.182521 5.3261380 4.9370546  2.6480839 4.9296906  0.000000 7.262477
#> 7   7.784517 1.9372094 2.3486714  5.7624556 2.8540771  7.262477 0.000000
#> 8   8.102018 2.9176827 2.8940312  6.6449346 2.3713446  7.266924 2.477155
#> 9  10.295279 4.7974208 4.9151412  8.7106610 4.5663701  9.481064 3.513262
#> 10  9.189900 3.3702736 3.6447061  7.3247977 3.7091975  8.537187 1.698193
#> 11  1.689399 5.5348340 5.3479688  1.5725896 5.8445798  2.592418 7.330633
#> 12  4.684898 1.2036179 1.0423882  2.7761567 1.8934405  4.224959 3.099956
#> 13  2.619555 3.3738261 3.1480655  0.7538375 3.6482197  2.502316 5.241617
#> 14  1.866263 4.7511530 4.5643210  0.7945045 5.0812537  2.428209 6.556950
#> 15 11.589801 5.8408992 6.0708162  9.8046156 5.9379914 10.860819 4.164835
#> 16  1.665573 4.2848168 4.0167181  0.9036885 4.3812527  1.744538 6.177357
#> 17  9.910939 4.0980465 4.3670383  8.0519230 4.3811502  9.243589 2.385423
#> 18  7.064095 1.8744591 1.7996333  5.5464563 1.3510748  6.280460 2.069315
#> 19  6.549551 1.8274773 1.5875660  5.1656634 0.8220168  5.724631 2.541307
#> 20  4.490602 1.7172308 1.7114929  2.3318837 2.6298557  4.217406 3.431692
#> 21  3.229411 2.7414095 2.3574563  2.0804164 2.5299685  2.585514 4.677103
#> 22  6.320921 1.0740194 0.9409086  4.7063686 0.8420695  5.607481 2.015904
#> 23 11.843960 6.1181587 6.3367877 10.0828568 6.1718604 11.099607 4.463916
#> 24  6.069598 0.7081536 0.5826479  4.3752541 0.9255663  5.406493 2.006773
#> 25  3.733868 2.1547828 1.9180057  1.8941919 2.5070352  3.337363 4.054026
#>           8         9        10         11        12        13         14
#> 1  8.102018 10.295279 9.1899001  1.6893992 4.6848979 2.6195548  1.8662626
#> 2  2.917683  4.797421 3.3702736  5.5348340 1.2036179 3.3738261  4.7511530
#> 3  2.894031  4.915141 3.6447061  5.3479688 1.0423882 3.1480655  4.5643210
#> 4  6.644935  8.710661 7.3247977  1.5725896 2.7761567 0.7538375  0.7945045
#> 5  2.371345  4.566370 3.7091975  5.8445798 1.8934405 3.6482197  5.0812537
#> 6  7.266924  9.481064 8.5371866  2.5924176 4.2249588 2.5023160  2.4282087
#> 7  2.477155  3.513262 1.6981930  7.3306327 3.0999556 5.2416166  6.5569501
#> 8  0.000000  2.216719 2.0218432  8.1636142 3.9352964 5.9524930  7.3882163
#> 9  2.216719  0.000000 1.9282166 10.2602497 5.9423801 8.0529943  9.4775972
#> 10 2.021843  1.928217 0.0000000  8.8971786 4.5735864 6.7436867  8.1151389
#> 11 8.163614 10.260250 8.8971786  0.0000000 4.3410572 2.2118329  0.7845378
#> 12 3.935296  5.942380 4.5735864  4.3410572 0.0000000 2.1702239  3.5565879
#> 13 5.952493  8.052994 6.7436867  2.2118329 2.1702239 0.0000000  1.4365452
#> 14 7.388216  9.477597 8.1151389  0.7845378 3.5565879 1.4365452  0.0000000
#> 15 3.715067  1.672027 2.4937866 11.3755745 7.0392595 9.2043371 10.5920369
#> 16 6.734901  8.880183 7.6457605  1.5443103 3.0856825 0.9572441  0.9329781
#> 17 2.466832  1.583544 0.7278135  9.6243866 5.3012918 7.4713305  8.8425418
#> 18 1.100222  3.233094 2.4431106  7.0706434 2.8384745 4.8588518  6.2933994
#> 19 1.552518  3.757995 3.0594808  6.6551975 2.5671151 4.4518240  5.8875168
#> 20 4.588137  6.514377 5.0201331  3.8989519 0.7382175 1.8749195  3.1262859
#> 21 4.898581  7.075288 5.9849644  3.3878578 1.6868006 1.3324172  2.6655108
#> 22 1.954849  4.023113 2.9414742  6.2431650 1.9806274 4.0328210  5.4625996
#> 23 3.913114  1.792189 2.7849745 11.6527553 7.3137952 9.4758248 10.8687804
#> 24 2.317031  4.337041 3.1338646  5.9235053 1.6241087 3.7187813  5.1406015
#> 25 4.750807  6.828253 5.5202570  3.4320065 0.9559350 1.2309404  2.6495934
#>            15         16        17        18        19        20        21
#> 1  11.5898014  1.6655729 9.9109390 7.0640947 6.5495510 4.4906022 3.2294115
#> 2   5.8408992  4.2848168 4.0980465 1.8744591 1.8274773 1.7172308 2.7414095
#> 3   6.0708162  4.0167181 4.3670383 1.7996333 1.5875660 1.7114929 2.3574563
#> 4   9.8046156  0.9036885 8.0519230 5.5464563 5.1656634 2.3318837 2.0804164
#> 5   5.9379914  4.3812527 4.3811502 1.3510748 0.8220168 2.6298557 2.5299685
#> 6  10.8608186  1.7445377 9.2435893 6.2804600 5.7246312 4.2174058 2.5855135
#> 7   4.1648352  6.1773572 2.3854227 2.0693153 2.5413071 3.4316920 4.6771030
#> 8   3.7150673  6.7349006 2.4668322 1.1002219 1.5525178 4.5881372 4.8985808
#> 9   1.6720269  8.8801829 1.5835443 3.2330941 3.7579947 6.5143769 7.0752883
#> 10  2.4937866  7.6457605 0.7278135 2.4431106 3.0594808 5.0201331 5.9849644
#> 11 11.3755745  1.5443103 9.6243866 7.0706434 6.6551975 3.8989519 3.3878578
#> 12  7.0392595  3.0856825 5.3012918 2.8384745 2.5671151 0.7382175 1.6868006
#> 13  9.2043371  0.9572441 7.4713305 4.8588518 4.4518240 1.8749195 1.3324172
#> 14 10.5920369  0.9329781 8.8425418 6.2933994 5.8875168 3.1262859 2.6655108
#> 15  0.0000000 10.0875001 1.7804148 4.5888798 5.1732365 7.5105095 8.3629172
#> 16 10.0875001  0.0000000 8.3720676 5.6573040 5.2010302 2.8320566 1.8727243
#> 17  1.7804148  8.3720676 0.0000000 3.0755256 3.6898568 5.7444981 6.7004239
#> 18  4.5888798  5.6573040 3.0755256 0.0000000 0.6166849 3.5036770 3.8421947
#> 19  5.1732365  5.2010302 3.6898568 0.6166849 0.0000000 3.2802277 3.3519792
#> 20  7.5105095  2.8320566 5.7444981 3.5036770 3.2802277 0.0000000 1.8991439
#> 21  8.3629172  1.8727243 6.7004239 3.8421947 3.3519792 1.8991439 0.0000000
#> 22  5.2715209  4.8603616 3.6384855 0.8587247 0.7702290 2.6472647 3.0921670
#> 23  0.3104311 10.3529895 2.0784940 4.8209472 5.3964416 7.7952708 8.6152991
#> 24  5.5247407  4.5712023 3.8477475 1.2299463 1.1193506 2.2740596 2.8527011
#> 25  7.9750688  2.1304590 6.2474199 3.6522792 3.2828784 0.9304418 0.9805912
#>           22         23        24        25
#> 1  6.3209208 11.8439597 6.0695981 3.7338678
#> 2  1.0740194  6.1181587 0.7081536 2.1547828
#> 3  0.9409086  6.3367877 0.5826479 1.9180057
#> 4  4.7063686 10.0828568 4.3752541 1.8941919
#> 5  0.8420695  6.1718604 0.9255663 2.5070352
#> 6  5.6074814 11.0996066 5.4064933 3.3373631
#> 7  2.0159040  4.4639164 2.0067725 4.0540263
#> 8  1.9548486  3.9131145 2.3170312 4.7508073
#> 9  4.0231133  1.7921887 4.3370414 6.8282533
#> 10 2.9414742  2.7849745 3.1338646 5.5202570
#> 11 6.2431650 11.6527553 5.9235053 3.4320065
#> 12 1.9806274  7.3137952 1.6241087 0.9559350
#> 13 4.0328210  9.4758248 3.7187813 1.2309404
#> 14 5.4625996 10.8687804 5.1406015 2.6495934
#> 15 5.2715209  0.3104311 5.5247407 7.9750688
#> 16 4.8603616 10.3529895 4.5712023 2.1304590
#> 17 3.6384855  2.0784940 3.8477475 6.2474199
#> 18 0.8587247  4.8209472 1.2299463 3.6522792
#> 19 0.7702290  5.3964416 1.1193506 3.2828784
#> 20 2.6472647  7.7952708 2.2740596 0.9304418
#> 21 3.0921670  8.6152991 2.8527011 0.9805912
#> 22 0.0000000  5.5231628 0.3807575 2.8144692
#> 23 5.5231628  0.0000000 5.7852808 8.2456269
#> 24 0.3807575  5.7852808 0.0000000 2.4915827
#> 25 2.8144692  8.2456269 2.4915827 0.0000000