Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3         4         5         6         7
#> 1  0.0000000 3.7094101 0.6806730 3.6984890 3.1446446 4.7799156 3.6295590
#> 2  3.7094101 0.0000000 3.0467750 0.0206010 0.8382880 1.1694767 0.4673952
#> 3  0.6806730 3.0467750 0.0000000 3.0350643 2.4642901 4.1029595 2.9949081
#> 4  3.6984890 0.0206010 3.0350643 0.0000000 0.8184250 1.1714517 0.4820427
#> 5  3.1446446 0.8382880 2.4642901 0.8184250 0.0000000 1.6574740 1.1419219
#> 6  4.7799156 1.1694767 4.1029595 1.1714517 1.6574740 0.0000000 1.5156047
#> 7  3.6295590 0.4673952 2.9949081 0.4820427 1.1419219 1.5156047 0.0000000
#> 8  0.7001196 3.7318141 0.8356529 3.7178105 3.0521176 4.7093293 3.7396845
#> 9  2.7187848 1.4146554 2.0436572 1.3954082 0.5859111 2.1921992 1.6586793
#> 10 2.4585650 2.9429388 2.2486571 2.9463226 2.9754289 4.0865643 2.5750349
#> 11 2.0015160 2.3265685 1.6016018 2.3253537 2.2038158 3.4958242 2.0440378
#> 12 1.3467521 2.4057640 0.6661001 2.3929557 1.7984399 3.4409668 2.3947572
#> 13 6.4653946 2.9587895 5.7847725 2.9601650 3.3253179 1.7898653 3.2895668
#> 14 3.3281097 6.7581653 3.9728843 6.7522295 6.3535230 7.9064370 6.5463084
#> 15 2.3421872 1.6556690 1.6672429 1.6381559 0.8837232 2.5332039 1.8176520
#> 16 5.5016883 1.7924405 4.8369979 1.8032015 2.4681906 0.9050764 1.9612987
#> 17 2.2796721 5.9715426 2.9253488 5.9595782 5.3484383 7.0036321 5.9085107
#> 18 2.4466279 1.6759705 1.9123379 1.6759236 1.6590203 2.8433073 1.3849023
#> 19 1.7227525 2.4460662 1.3287457 2.4426699 2.2266350 3.6128496 2.2052156
#> 20 6.3051769 2.6471141 5.6621715 2.6622014 3.4107392 1.9413708 2.6770424
#> 21 1.9226858 5.6284672 2.5989151 5.6179873 5.0606389 6.7017047 5.5268068
#> 22 3.7318936 7.1718762 4.3811832 7.1660056 6.7673497 8.3206613 6.9578266
#> 23 9.2151424 5.5061738 8.5522681 5.5175088 6.1589954 4.5105985 5.6240912
#> 24 1.9079095 5.4401675 2.5624025 5.4325190 4.9797377 6.5679651 5.2699456
#> 25 1.2363504 2.4764969 0.6079377 2.4661339 1.9670254 3.5683135 2.3971257
#>            8         9        10        11        12        13         14
#> 1  0.7001196 2.7187848 2.4585650 2.0015160 1.3467521  6.465395  3.3281097
#> 2  3.7318141 1.4146554 2.9429388 2.3265685 2.4057640  2.958789  6.7581653
#> 3  0.8356529 2.0436572 2.2486571 1.6016018 0.6661001  5.784773  3.9728843
#> 4  3.7178105 1.3954082 2.9463226 2.3253537 2.3929557  2.960165  6.7522295
#> 5  3.0521176 0.5859111 2.9754289 2.2038158 1.7984399  3.325318  6.3535230
#> 6  4.7093293 2.1921992 4.0865643 3.4958242 3.4409668  1.789865  7.9064370
#> 7  3.7396845 1.6586793 2.5750349 2.0440378 2.3947572  3.289567  6.5463084
#> 8  0.0000000 2.5424290 3.0205708 2.4296578 1.3450224  6.306116  3.7824085
#> 9  2.5424290 0.0000000 3.0007905 2.1674595 1.3899048  3.769554  6.0013470
#> 10 3.0205708 3.0007905 0.0000000 0.8514616 2.2256907  5.864455  4.3136794
#> 11 2.4296578 2.1674595 0.8514616 0.0000000 1.4250942  5.285358  4.5243114
#> 12 1.3450224 1.3899048 2.2256907 1.4250942 0.0000000  5.118891  4.6119373
#> 13 6.3061156 3.7695543 5.8644554 5.2853580 5.1188908  0.000000  9.6624302
#> 14 3.7824085 6.0013470 4.3136794 4.5243114 4.6119373  9.662430  0.0000000
#> 15 2.1813200 0.3766103 2.7747590 1.9263725 1.0161304  4.140742  5.6280385
#> 16 5.4989059 3.0343893 4.4609590 4.0023407 4.1895823  1.601783  8.5061037
#> 17 2.3201618 4.8580547 4.4043421 4.1827702 3.5714805  8.625299  2.2146624
#> 18 2.7398302 1.7589223 1.3218780 0.6595977 1.5017684  4.633146  5.1715050
#> 19 2.1530443 2.1155698 1.0148438 0.2803378 1.2116468  5.397274  4.3414151
#> 20 6.3751697 3.9920365 4.9333594 4.6243415 5.0396670  2.088553  9.1451365
#> 21 2.1347835 4.6080844 3.8645192 3.7100161 3.2630046  8.372607  1.8666591
#> 22 4.1685767 6.4124928 4.7033163 4.9328052 5.0227894 10.076787  0.4143822
#> 23 9.2064713 6.7027896 7.9280833 7.6154600 7.9046608  3.280107 12.1393163
#> 24 2.3721288 4.6012877 3.2944208 3.3119997 3.2119945  8.301835  1.4227659
#> 25 1.4199677 1.6364237 1.8902116 1.1215383 0.3572918  5.288741  4.3895490
#>           15        16         17        18        19       20         21
#> 1  2.3421872 5.5016883  2.2796721 2.4466279 1.7227525 6.305177  1.9226858
#> 2  1.6556690 1.7924405  5.9715426 1.6759705 2.4460662 2.647114  5.6284672
#> 3  1.6672429 4.8369979  2.9253488 1.9123379 1.3287457 5.662171  2.5989151
#> 4  1.6381559 1.8032015  5.9595782 1.6759236 2.4426699 2.662201  5.6179873
#> 5  0.8837232 2.4681906  5.3484383 1.6590203 2.2266350 3.410739  5.0606389
#> 6  2.5332039 0.9050764  7.0036321 2.8433073 3.6128496 1.941371  6.7017047
#> 7  1.8176520 1.9612987  5.9085107 1.3849023 2.2052156 2.677042  5.5268068
#> 8  2.1813200 5.4989059  2.3201618 2.7398302 2.1530443 6.375170  2.1347835
#> 9  0.3766103 3.0343893  4.8580547 1.7589223 2.1155698 3.992037  4.6080844
#> 10 2.7747590 4.4609590  4.4043421 1.3218780 1.0148438 4.933359  3.8645192
#> 11 1.9263725 4.0023407  4.1827702 0.6595977 0.2803378 4.624341  3.7100161
#> 12 1.0161304 4.1895823  3.5714805 1.5017684 1.2116468 5.039667  3.2630046
#> 13 4.1407421 1.6017829  8.6252987 4.6331463 5.3972741 2.088553  8.3726070
#> 14 5.6280385 8.5061037  2.2146624 5.1715050 4.3414151 9.145137  1.8666591
#> 15 0.0000000 3.3515667  4.4917056 1.6135145 1.8376634 4.282360  4.2338840
#> 16 3.3515667 0.0000000  7.7599643 3.3446056 4.1660247 1.041890  7.4208469
#> 17 4.4917056 7.7599643  0.0000000 4.7002581 3.9184033 8.582728  0.6042244
#> 18 1.6135145 3.3446056  4.7002581 0.0000000 0.8363876 3.991169  4.2638560
#> 19 1.8376634 4.1660247  3.9184033 0.8363876 0.0000000 4.827504  3.4563524
#> 20 4.2823597 1.0418904  8.5827284 3.9911695 4.8275043 0.000000  8.2032106
#> 21 4.2338840 7.4208469  0.6042244 4.2638560 3.4563524 8.203211  0.0000000
#> 22 6.0387577 8.9180536  2.4663460 5.5815783 4.7526839 9.551268  2.1889120
#> 23 7.0404252 3.7154882 11.4754227 6.9741620 7.8097230 2.999405 11.1317218
#> 24 4.2259930 7.2148076  1.4588678 3.9278247 3.0923309 7.918181  0.8671726
#> 25 1.2782728 4.2688353  3.5122713 1.3219351 0.8840432 5.070483  3.1520162
#>            22        23         24        25
#> 1   3.7318936  9.215142  1.9079095 1.2363504
#> 2   7.1718762  5.506174  5.4401675 2.4764969
#> 3   4.3811832  8.552268  2.5624025 0.6079377
#> 4   7.1660056  5.517509  5.4325190 2.4661339
#> 5   6.7673497  6.158995  4.9797377 1.9670254
#> 6   8.3206613  4.510599  6.5679651 3.5683135
#> 7   6.9578266  5.624091  5.2699456 2.3971257
#> 8   4.1685767  9.206471  2.3721288 1.4199677
#> 9   6.4124928  6.702790  4.6012877 1.6364237
#> 10  4.7033163  7.928083  3.2944208 1.8902116
#> 11  4.9328052  7.615460  3.3119997 1.1215383
#> 12  5.0227894  7.904661  3.2119945 0.3572918
#> 13 10.0767869  3.280107  8.3018353 5.2887413
#> 14  0.4143822 12.139316  1.4227659 4.3895490
#> 15  6.0387577  7.040425  4.2259930 1.2782728
#> 16  8.9180536  3.715488  7.2148076 4.2688353
#> 17  2.4663460 11.475423  1.4588678 3.5122713
#> 18  5.5815783  6.974162  3.9278247 1.3219351
#> 19  4.7526839  7.809723  3.0923309 0.8840432
#> 20  9.5512675  2.999405  7.9181805 5.0704831
#> 21  2.1889120 11.131722  0.8671726 3.1520162
#> 22  0.0000000 12.546670  1.8240931 4.8028896
#> 23 12.5466701  0.000000 10.8922436 7.9805050
#> 24  1.8240931 10.892244  0.0000000 3.0133313
#> 25  4.8028896  7.980505  3.0133313 0.0000000