Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>           1          2         3          4         5        6          7
#> 1  0.000000  6.0899274  8.877672  6.5687639 3.2556661 3.613887  6.0772460
#> 2  6.089927  0.0000000  2.800118  1.0746388 3.4256298 4.117911  1.5076246
#> 3  8.877672  2.8001177  0.000000  2.7178116 6.1570243 6.710265  3.4887145
#> 4  6.568764  1.0746388  2.717812  0.0000000 3.5610847 4.011656  0.7814609
#> 5  3.255666  3.4256298  6.157024  3.5610847 0.0000000 1.025532  2.9413988
#> 6  3.613887  4.1179114  6.710265  4.0116560 1.0255324 0.000000  3.2854921
#> 7  6.077246  1.5076246  3.488715  0.7814609 2.9413988 3.285492  0.0000000
#> 8  2.487489  6.8597849  9.595674  6.9606447 3.4409328 3.036972  6.2771140
#> 9  4.426887  1.9249811  4.550440  2.8110904 2.6036526 3.585851  2.7828972
#> 10 3.755413  2.8215698  5.542365  2.9525913 0.6151539 1.412003  2.3566097
#> 11 1.790186  7.5453952 10.343854  7.8606836 4.3234421 4.268440  7.2647689
#> 12 3.981455  2.1750303  4.970630  2.5969706 1.3787321 2.300900  2.2057961
#> 13 3.865676  2.5803299  5.320891  2.7677965 0.8453037 1.673031  2.2141274
#> 14 7.219455  1.7649078  2.538368  0.7447798 4.1178344 4.436718  1.1780953
#> 15 5.778629  0.3575453  3.134244  1.0928012 3.0695963 3.762939  1.3280432
#> 16 2.545521  3.9306889  6.709146  4.1850213 0.7101465 1.366048  3.6069284
#> 17 6.356159  0.8144452  2.606727  1.7907244 3.9899863 4.776069  2.3170461
#> 18 7.014870  1.5461927  2.574217  0.5139433 3.9374115 4.294146  1.0126532
#> 19 6.424871  0.7637221  2.687441  0.3111800 3.4991367 4.023991  0.9329662
#> 20 3.504878  2.8788499  5.643562  3.1200062 0.5959467 1.562355  2.5753088
#> 21 6.099537 11.8612888 14.645336 12.0703080 8.5129632 8.184538 11.4099050
#> 22 4.347193 10.2083377 13.004077 10.4826253 6.9239501 6.709425  9.8546198
#> 23 4.175039  1.9207119  4.702781  2.5758132 1.9323246 2.886956  2.3596493
#> 24 2.427148  6.7536136  9.490770  6.8578363 3.3354867 2.941513  6.1761987
#> 25 5.217247  1.4154371  3.823030  2.4471943 3.2312283 4.156237  2.6373630
#>            8          9        10        11        12        13         14
#> 1  2.4874889  4.4268874 3.7554126  1.790186 3.9814552 3.8656764  7.2194549
#> 2  6.8597849  1.9249811 2.8215698  7.545395 2.1750303 2.5803299  1.7649078
#> 3  9.5956742  4.5504400 5.5423652 10.343854 4.9706302 5.3208906  2.5383678
#> 4  6.9606447  2.8110904 2.9525913  7.860684 2.5969706 2.7677965  0.7447798
#> 5  3.4409328  2.6036526 0.6151539  4.323442 1.3787321 0.8453037  4.1178344
#> 6  3.0369716  3.5858511 1.4120027  4.268440 2.3008996 1.6730311  4.4367182
#> 7  6.2771140  2.7828972 2.3566097  7.264769 2.2057961 2.2141274  1.1780953
#> 8  0.0000000  5.7051050 4.0542665  1.683488 4.7296510 4.2833971  7.4463106
#> 9  5.7051050  0.0000000 2.2009165  6.049116 1.3137872 1.9578080  3.5557895
#> 10 4.0542665  2.2009165 0.0000000  4.916006 0.8949674 0.2624100  3.5275152
#> 11 1.6834877  6.0491157 4.9160064  0.000000 5.3735423 5.0930008  8.4405057
#> 12 4.7296510  1.3137872 0.8949674  5.373542 0.0000000 0.6449802  3.2789485
#> 13 4.2833971  1.9578080 0.2624100  5.093001 0.6449802 0.0000000  3.3715858
#> 14 7.4463106  3.5557895 3.5275152  8.440506 3.2789485 3.3715858  0.0000000
#> 15 6.5055004  1.7386929 2.4645291  7.210024 1.8365104 2.2243874  1.8313173
#> 16 2.9674529  2.7857441 1.2503618  3.675683 1.7699887 1.4173332  4.7764628
#> 17 7.3742352  1.9581658 3.4141777  7.915607 2.6454713 3.1577419  2.3786396
#> 18 7.2896874  3.3249977 3.3409949  8.256790 3.0637236 3.1769887  0.2308478
#> 19 6.9251569  2.5452939 2.8848281  7.765680 2.4434242 2.6812001  1.0249566
#> 20 3.9833120  2.0259949 0.3266841  4.743768 0.7833523 0.3612143  3.7313306
#> 21 5.1479873 10.4273620 9.1263079  4.379843 9.6877291 9.3334719 12.5839331
#> 22 3.7104849  8.7117162 7.5300427  2.670465 8.0340251 7.7223167 11.0326747
#> 23 5.1788864  0.7211496 1.4881070  5.685548 0.5950892 1.2400693  3.3043418
#> 24 0.1070032  5.5989260 3.9490544  1.706522 4.6227667 4.1776279  7.3461584
#> 25 6.4537052  0.7921388 2.7443535  6.839356 1.8610605 2.4843085  3.1687549
#>            15        16         17         18         19        20        21
#> 1   5.7786293 2.5455212  6.3561593  7.0148705  6.4248707 3.5048783  6.099537
#> 2   0.3575453 3.9306889  0.8144452  1.5461927  0.7637221 2.8788499 11.861289
#> 3   3.1342437 6.7091462  2.6067270  2.5742173  2.6874415 5.6435618 14.645336
#> 4   1.0928012 4.1850213  1.7907244  0.5139433  0.3111800 3.1200062 12.070308
#> 5   3.0695963 0.7101465  3.9899863  3.9374115  3.4991367 0.5959467  8.512963
#> 6   3.7629390 1.3660482  4.7760694  4.2941456  4.0239905 1.5623549  8.184538
#> 7   1.3280432 3.6069284  2.3170461  1.0126532  0.9329662 2.5753088 11.409905
#> 8   6.5055004 2.9674529  7.3742352  7.2896874  6.9251569 3.9833120  5.147987
#> 9   1.7386929 2.7857441  1.9581658  3.3249977  2.5452939 2.0259949 10.427362
#> 10  2.4645291 1.2503618  3.4141777  3.3409949  2.8848281 0.3266841  9.126308
#> 11  7.2100240 3.6756829  7.9156071  8.2567902  7.7656803 4.7437678  4.379843
#> 12  1.8365104 1.7699887  2.6454713  3.0637236  2.4434242 0.7833523  9.687729
#> 13  2.2243874 1.4173332  3.1577419  3.1769887  2.6812001 0.3612143  9.333472
#> 14  1.8313173 4.7764628  2.3786396  0.2308478  1.0249566 3.7313306 12.583933
#> 15  0.0000000 3.5828887  1.0939042  1.6020164  0.8098789 2.5266814 11.516787
#> 16  3.5828887 0.0000000  4.4094333  4.5872542  4.0943883 1.0714402  7.936200
#> 17  1.0939042 4.4094333  0.0000000  2.1873102  1.4908344 3.4099747 12.274798
#> 18  1.6020164 4.5872542  2.1873102  0.0000000  0.7980653 3.5353424 12.420871
#> 19  0.8098789 4.0943883  1.4908344  0.7980653  0.0000000 3.0233191 12.010844
#> 20  2.5266814 1.0714402  3.4099747  3.5353424  3.0233191 0.0000000  9.003080
#> 21 11.5167871 7.9362000 12.2747983 12.4208713 12.0108435 9.0030795  0.000000
#> 22  9.8699230 6.3092991 10.5860290 10.8584412 10.4029423 7.3807377  1.764136
#> 23  1.6298114 2.2155877  2.2343576  3.0773339  2.3626937 1.3407610 10.040563
#> 24  6.3994296 2.8604501  7.2672985  7.1886909  6.8211216 3.8768739  5.245007
#> 25  1.3679170 3.5049565  1.2164627  2.9443200  2.1469197 2.6362557 11.216617
#>           22         23        24         25
#> 1   4.347193  4.1750391 2.4271484  5.2172468
#> 2  10.208338  1.9207119 6.7536136  1.4154371
#> 3  13.004077  4.7027806 9.4907705  3.8230298
#> 4  10.482625  2.5758132 6.8578363  2.4471943
#> 5   6.923950  1.9323246 3.3354867  3.2312283
#> 6   6.709425  2.8869559 2.9415129  4.1562366
#> 7   9.854620  2.3596493 6.1761987  2.6373630
#> 8   3.710485  5.1788864 0.1070032  6.4537052
#> 9   8.711716  0.7211496 5.5989260  0.7921388
#> 10  7.530043  1.4881070 3.9490544  2.7443535
#> 11  2.670465  5.6855476 1.7065219  6.8393560
#> 12  8.034025  0.5950892 4.6227667  1.8610605
#> 13  7.722317  1.2400693 4.1776279  2.4843085
#> 14 11.032675  3.3043418 7.3461584  3.1687549
#> 15  9.869923  1.6298114 6.3994296  1.3679170
#> 16  6.309299  2.2155877 2.8604501  3.5049565
#> 17 10.586029  2.2343576 7.2672985  1.2164627
#> 18 10.858441  3.0773339 7.1886909  2.9443200
#> 19 10.402942  2.3626937 6.8211216  2.1469197
#> 20  7.380738  1.3407610 3.8768739  2.6362557
#> 21  1.764136 10.0405630 5.2450067 11.2166173
#> 22  0.000000  8.3558319 3.7946584  9.5031059
#> 23  8.355832  0.0000000 5.0719401  1.3047412
#> 24  3.794658  5.0719401 0.0000000  6.3470530
#> 25  9.503106  1.3047412 6.3470530  0.0000000