Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3        4         5         6         7
#> 1  0.0000000 7.5858526 7.8653887 2.346773 4.6014973 0.4288087 4.2260203
#> 2  7.5858526 0.0000000 0.5161222 5.453044 3.2690195 7.2097706 3.3598328
#> 3  7.8653887 0.5161222 0.0000000 5.797642 3.4202404 7.5016613 3.6539267
#> 4  2.3467730 5.4530436 5.7976416 0.000000 2.9488775 1.9221822 2.1917120
#> 5  4.6014973 3.2690195 3.4202404 2.948877 0.0000000 4.2865396 1.0657039
#> 6  0.4288087 7.2097706 7.5016613 1.922182 4.2865396 0.0000000 3.8523017
#> 7  4.2260203 3.3598328 3.6539267 2.191712 1.0657039 3.8523017 0.0000000
#> 8  4.5779204 3.0165547 3.3363417 2.465894 1.2128788 4.1956848 0.3901945
#> 9  7.5483839 0.8005815 0.4962441 5.553102 3.0312562 7.1968058 3.3759050
#> 10 5.9408573 1.8357707 1.9755783 4.011584 1.4471058 5.5944789 1.8200910
#> 11 1.3639281 7.9007384 8.2485671 2.450959 5.2538960 1.2856413 4.6255664
#> 12 8.0752373 1.2205767 0.7227723 6.125142 3.5106522 7.7329732 3.9390882
#> 13 7.7067411 2.4727322 2.9497169 5.362153 4.4866196 7.2839448 3.9378044
#> 14 7.2358391 0.6987669 0.6582345 5.212287 2.7631950 6.8784831 3.0437679
#> 15 4.5882846 3.0304510 3.2791558 2.652966 0.6924911 4.2318188 0.4926332
#> 16 8.7863448 1.3414905 1.4163052 6.569308 4.5952397 8.3955407 4.5819571
#> 17 3.3179144 4.2688585 4.5630870 1.359769 1.5974554 2.9419559 0.9104184
#> 18 1.2722555 6.3735437 6.6783704 1.083101 3.5443169 0.8569783 3.0244634
#> 19 3.5436818 4.0531894 4.3217135 1.692592 1.2577443 3.1847621 0.7182503
#> 20 4.3188133 3.4502596 3.8334454 2.041716 1.8818331 3.9102767 0.8317383
#> 21 7.0588581 1.2208942 1.7369143 4.788541 3.2556929 6.6556407 2.9491551
#> 22 4.0013312 3.7373496 4.1065861 1.732735 1.8814212 3.5932209 0.8195157
#> 23 5.8282350 1.8165734 2.0390178 3.821633 1.4719839 5.4688058 1.6388026
#> 24 0.6962043 8.1986507 8.4591530 3.036497 5.1361985 1.1249053 4.8439173
#> 25 2.1807893 6.3850179 6.7669644 1.135377 4.0768077 1.7886460 3.2580021
#>            8         9        10       11        12       13        14
#> 1  4.5779204 7.5483839 5.9408573 1.363928 8.0752373 7.706741 7.2358391
#> 2  3.0165547 0.8005815 1.8357707 7.900738 1.2205767 2.472732 0.6987669
#> 3  3.3363417 0.4962441 1.9755783 8.248567 0.7227723 2.949717 0.6582345
#> 4  2.4658942 5.5531024 4.0115843 2.450959 6.1251418 5.362153 5.2122873
#> 5  1.2128788 3.0312562 1.4471058 5.253896 3.5106522 4.486620 2.7631950
#> 6  4.1956848 7.1968058 5.5944789 1.285641 7.7329732 7.283945 6.8784831
#> 7  0.3901945 3.3759050 1.8200910 4.625566 3.9390882 3.937804 3.0437679
#> 8  0.0000000 3.0891802 1.5979875 4.915159 3.6671390 3.556806 2.7465612
#> 9  3.0891802 0.0000000 1.6109743 7.998299 0.6018092 3.259545 0.3596477
#> 10 1.5979875 1.6109743 0.0000000 6.437499 2.1395811 3.474413 1.3210183
#> 11 4.9151586 7.9982989 6.4374995 0.000000 8.5646532 7.584763 7.6601464
#> 12 3.6671390 0.6018092 2.1395811 8.564653 0.0000000 3.672472 0.9614215
#> 13 3.5568057 3.2595445 3.4744133 7.584763 3.6724719 0.000000 3.0482895
#> 14 2.7465612 0.3596477 1.3210183 7.660146 0.9614215 3.048289 0.0000000
#> 15 0.5230320 2.9654466 1.3737058 5.066122 3.5110703 3.912181 2.6477932
#> 16 4.2141638 1.9114852 3.1735624 8.989252 1.9527805 2.326406 1.9820338
#> 17 1.2626638 4.2752493 2.6945810 3.745623 4.8293033 4.652880 3.9480087
#> 18 3.3570184 6.3885052 4.7966154 1.735096 6.9358082 6.435880 6.0635059
#> 19 1.1045124 4.0124261 2.4150046 4.046641 4.5544980 4.611831 3.6941637
#> 20 0.7375093 3.6523176 2.2727170 4.466430 4.2469700 3.470220 3.2970774
#> 21 2.5611958 1.9294136 2.0822465 7.183577 2.4315517 1.434824 1.6658694
#> 22 0.8740099 3.9044244 2.4663365 4.167592 4.4939079 3.777259 3.5526628
#> 23 1.3684435 1.7372671 0.3282875 6.263041 2.3035167 3.223969 1.4099814
#> 24 5.2061785 8.1244820 6.5135739 1.740988 8.6362434 8.391272 7.8212256
#> 25 3.4845674 6.5624236 5.0689583 1.646817 7.1476026 5.939682 6.2126481
#>           15       16        17        18        19        20       21
#> 1  4.5882846 8.786345 3.3179144 1.2722555 3.5436818 4.3188133 7.058858
#> 2  3.0304510 1.341490 4.2688585 6.3735437 4.0531894 3.4502596 1.220894
#> 3  3.2791558 1.416305 4.5630870 6.6783704 4.3217135 3.8334454 1.736914
#> 4  2.6529662 6.569308 1.3597688 1.0831005 1.6925915 2.0417158 4.788541
#> 5  0.6924911 4.595240 1.5974554 3.5443169 1.2577443 1.8818331 3.255693
#> 6  4.2318188 8.395541 2.9419559 0.8569783 3.1847621 3.9102767 6.655641
#> 7  0.4926332 4.581957 0.9104184 3.0244634 0.7182503 0.8317383 2.949155
#> 8  0.5230320 4.214164 1.2626638 3.3570184 1.1045124 0.7375093 2.561196
#> 9  2.9654466 1.911485 4.2752493 6.3885052 4.0124261 3.6523176 1.929414
#> 10 1.3737058 3.173562 2.6945810 4.7966154 2.4150046 2.2727170 2.082246
#> 11 5.0661217 8.989252 3.7456228 1.7350963 4.0466415 4.4664299 7.183577
#> 12 3.5110703 1.952780 4.8293033 6.9358082 4.5544980 4.2469700 2.431552
#> 13 3.9121814 2.326406 4.6528795 6.4358798 4.6118314 3.4702203 1.434824
#> 14 2.6477932 1.982034 3.9480087 6.0635059 3.6941637 3.2970774 1.665869
#> 15 0.0000000 4.302911 1.3213030 3.4253040 1.0470853 1.2265007 2.784380
#> 16 4.3029108 0.000000 5.4766856 7.5461795 5.2932736 4.5289848 1.831549
#> 17 1.3213030 5.476686 0.0000000 2.1159350 0.3428864 1.2028382 3.795810
#> 18 3.4253040 7.546180 2.1159350 0.0000000 2.3818776 3.0534539 5.799887
#> 19 1.0470853 5.293274 0.3428864 2.3818776 0.0000000 1.2512347 3.665708
#> 20 1.2265007 4.528985 1.2028382 3.0534539 1.2512347 0.0000000 2.750541
#> 21 2.7843802 1.831549 3.7958104 5.7998868 3.6657080 2.7505414 0.000000
#> 22 1.2857523 4.836655 0.9332192 2.7365068 1.0442362 0.3175729 3.065531
#> 23 1.2403965 3.128423 2.5391022 4.6539901 2.2851445 2.0029231 1.868719
#> 24 5.1808220 9.420205 3.9435470 1.9561348 4.1457483 4.9841560 7.713151
#> 25 3.7372705 7.412852 2.4806865 1.3169872 2.8190643 2.9360339 5.591558
#>           22        23        24       25
#> 1  4.0013312 5.8282350 0.6962043 2.180789
#> 2  3.7373496 1.8165734 8.1986507 6.385018
#> 3  4.1065861 2.0390178 8.4591530 6.766964
#> 4  1.7327355 3.8216333 3.0364966 1.135377
#> 5  1.8814212 1.4719839 5.1361985 4.076808
#> 6  3.5932209 5.4688058 1.1249053 1.788646
#> 7  0.8195157 1.6388026 4.8439173 3.258002
#> 8  0.8740099 1.3684435 5.2061785 3.484567
#> 9  3.9044244 1.7372671 8.1244820 6.562424
#> 10 2.4663365 0.3282875 6.5135739 5.068958
#> 11 4.1675924 6.2630411 1.7409882 1.646817
#> 12 4.4939079 2.3035167 8.6362434 7.147603
#> 13 3.7772588 3.2239687 8.3912724 5.939682
#> 14 3.5526628 1.4099814 7.8212256 6.212648
#> 15 1.2857523 1.2403965 5.1808220 3.737270
#> 16 4.8366546 3.1284232 9.4202046 7.412852
#> 17 0.9332192 2.5391022 3.9435470 2.480687
#> 18 2.7365068 4.6539901 1.9561348 1.316987
#> 19 1.0442362 2.2851445 4.1457483 2.819064
#> 20 0.3175729 2.0029231 4.9841560 2.936034
#> 21 3.0655306 1.8687193 7.7131509 5.591558
#> 22 0.0000000 2.2184282 4.6666592 2.661884
#> 23 2.2184282 0.0000000 6.4202162 4.853010
#> 24 4.6666592 6.4202162 0.0000000 2.845495
#> 25 2.6618836 4.8530101 2.8454950 0.000000