Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1          2          3         4         5         6          7
#> 1  0.0000000  5.5160204  4.8688755 0.5588097 2.4660078 2.0880094  5.9806215
#> 2  5.5160204  0.0000000  0.9119437 4.9671266 6.6223567 3.6859786  1.7837481
#> 3  4.8688755  0.9119437  0.0000000 4.3110152 6.2517413 3.2374450  1.4645730
#> 4  0.5588097  4.9671266  4.3110152 0.0000000 2.7564877 1.6495728  5.4263960
#> 5  2.4660078  6.6223567  6.2517413 2.7564877 0.0000000 3.0203536  7.6076983
#> 6  2.0880094  3.6859786  3.2374450 1.6495728 3.0203536 0.0000000  4.5903628
#> 7  5.9806215  1.7837481  1.4645730 5.4263960 7.6076983 4.5903628  0.0000000
#> 8  2.1744743  5.8342013  5.4839218 2.3154453 0.7935018 2.2695936  6.8587047
#> 9  0.8008984  4.7273476  4.0686605 0.2423586 2.8962668 1.4744861  5.1886535
#> 10 2.5841289  3.2065674  2.8187435 2.1190193 3.4409405 0.5130164  4.2161701
#> 11 5.3937792  0.3080277  1.0421241 4.8514654 6.4048636 3.5062732  2.0780732
#> 12 6.1524609  0.8715836  1.2891186 5.5954558 7.4208797 4.4402853  1.2156035
#> 13 2.4289059  3.1253564  2.5946517 1.9048530 3.6952858 0.6927729  3.9129332
#> 14 4.3377811  1.8543116  1.9789115 3.8486659 4.9460668 2.2731474  3.4070956
#> 15 5.4667146 10.4668905 10.0262996 5.9777991 3.8772921 6.8004026 11.3064365
#> 16 3.3254731  2.1966112  1.5816790 2.7733054 4.7082828 1.6893171  2.9011102
#> 17 1.5523300  3.9670668  3.3531592 1.0172048 3.1848847 0.8191468  4.5584936
#> 18 7.0203497  2.2318088  2.3051021 6.4638249 8.5524527 5.5336903  1.0645417
#> 19 2.9575245  4.6413220  3.7431065 2.6089430 5.3641717 3.3210000  4.1753308
#> 20 6.4315575  1.6056317  1.6557578 5.8733133 7.9073101 4.8912644  0.6660971
#> 21 3.9382530  2.5706903  1.6588605 3.4012680 5.8556006 2.9837235  2.1491910
#> 22 5.8098112  1.9356487  1.4813246 5.2589595 7.5048297 4.4967813  0.2845837
#> 23 5.5544762  0.5839448  0.6924979 4.9973483 6.8603433 3.8647617  1.2038618
#> 24 7.7738559  3.4495220  3.3648228 7.2298102 9.5108389 6.4992660  1.9183049
#> 25 4.8165112  0.9773000  1.2401766 4.2880013 5.7112143 2.8583244  2.5852961
#>            8         9        10         11         12        13        14
#> 1  2.1744743 0.8008984 2.5841289  5.3937792  6.1524609 2.4289059 4.3377811
#> 2  5.8342013 4.7273476 3.2065674  0.3080277  0.8715836 3.1253564 1.8543116
#> 3  5.4839218 4.0686605 2.8187435  1.0421241  1.2891186 2.5946517 1.9789115
#> 4  2.3154453 0.2423586 2.1190193  4.8514654  5.5954558 1.9048530 3.8486659
#> 5  0.7935018 2.8962668 3.4409405  6.4048636  7.4208797 3.6952858 4.9460668
#> 6  2.2695936 1.4744861 0.5130164  3.5062732  4.4402853 0.6927729 2.2731474
#> 7  6.8587047 5.1886535 4.2161701  2.0780732  1.2156035 3.9129332 3.4070956
#> 8  0.0000000 2.4021480 2.6665144  5.6139409  6.6385507 2.9560923 4.1529804
#> 9  2.4021480 0.0000000 1.9239447  4.6142197  5.3531951 1.6788613 3.6356967
#> 10 2.6665144 1.9239447 0.0000000  3.0147461  3.9836803 0.5533140 1.7622287
#> 11 5.6139409 4.6142197 3.0147461  0.0000000  1.1670985 2.9794762 1.5741556
#> 12 6.6385507 5.3531951 3.9836803  1.1670985  0.0000000 3.8327537 2.7248107
#> 13 2.9560923 1.6788613 0.5533140  2.9794762  3.8327537 0.0000000 1.9785558
#> 14 4.1529804 3.6356967 1.7622287  1.5741556  2.7248107 1.9785558 0.0000000
#> 15 4.6698068 6.1965729 7.2623656 10.2616802 11.2401646 7.4323221 8.8226743
#> 16 3.9578160 2.5326285 1.3414676  2.1030549  2.8418485 1.0199997 1.6121149
#> 17 2.5391461 0.7895375 1.1827795  3.8414674  4.6259822 0.8893383 2.8526965
#> 18 7.7878512 6.2243229 5.1236720  2.5361604  1.3800487 4.8744040 4.0598722
#> 19 4.8951325 2.4939081 3.4656442  4.6918184  4.8898483 2.9253620 4.5324914
#> 20 7.1394912 5.6323095 4.4737211  1.9132296  0.8022874 4.2392461 3.4117564
#> 21 5.1756702 3.1754842 2.7899573  2.6708859  2.7611118 2.3077659 2.9880873
#> 22 6.7661150 5.0235309 4.1452457  2.2166530  1.4564757 3.8114819 3.4559021
#> 23 6.0826513 4.7550714 3.4194843  0.8742116  0.5982509 3.2455306 2.2950512
#> 24 8.7688306 6.9983155 6.1335170  3.7570059  2.6199167 5.8163520 5.2380766
#> 25 4.9189344 4.0567141 2.3551698  0.7086590  1.8467783 2.3876067 0.8782851
#>           15       16        17         18       19         20       21
#> 1   5.466715 3.325473 1.5523300  7.0203497 2.957524  6.4315575 3.938253
#> 2  10.466891 2.196611 3.9670668  2.2318088 4.641322  1.6056317 2.570690
#> 3  10.026300 1.581679 3.3531592  2.3051021 3.743106  1.6557578 1.658860
#> 4   5.977799 2.773305 1.0172048  6.4638249 2.608943  5.8733133 3.401268
#> 5   3.877292 4.708283 3.1848847  8.5524527 5.364172  7.9073101 5.855601
#> 6   6.800403 1.689317 0.8191468  5.5336903 3.321000  4.8912644 2.983723
#> 7  11.306436 2.901110 4.5584936  1.0645417 4.175331  0.6660971 2.149191
#> 8   4.669807 3.957816 2.5391461  7.7878512 4.895133  7.1394912 5.175670
#> 9   6.196573 2.532629 0.7895375  6.2243229 2.493908  5.6323095 3.175484
#> 10  7.262366 1.341468 1.1827795  5.1236720 3.465644  4.4737211 2.789957
#> 11 10.261680 2.103055 3.8414674  2.5361604 4.691818  1.9132296 2.670886
#> 12 11.240165 2.841849 4.6259822  1.3800487 4.889848  0.8022874 2.761112
#> 13  7.432322 1.020000 0.8893383  4.8744040 2.925362  4.2392461 2.307766
#> 14  8.822674 1.612115 2.8526965  4.0598722 4.532491  3.4117564 2.988087
#> 15  0.000000 8.450422 6.7515750 12.2996336 8.282938 11.6702370 9.369356
#> 16  8.450422 0.000000 1.7859536  3.8544164 2.924654  3.2200087 1.565928
#> 17  6.751575 1.785954 0.0000000  5.5692871 2.575734  4.9570773 2.675747
#> 18 12.299634 3.854416 5.5692871  0.0000000 5.215801  0.6582443 3.213588
#> 19  8.282938 2.924654 2.5757343  5.2158007 0.000000  4.7971683 2.128748
#> 20 11.670237 3.220009 4.9570773  0.6582443 4.797168  0.0000000 2.721252
#> 21  9.369356 1.565928 2.6757470  3.2135885 2.128748  2.7212517 0.000000
#> 22 11.169559 2.813369 4.4180007  1.2992129 3.919043  0.9505175 1.930925
#> 23 10.664687 2.247646 4.0299963  1.7747718 4.363140  1.1199795 2.241609
#> 24 13.166827 4.811931 6.4173782  1.2507149 5.585831  1.8439187 3.841368
#> 25  9.575474 1.639571 3.2708872  3.1856669 4.455582  2.5414470 2.607553
#>            22         23        24        25
#> 1   5.8098112  5.5544762  7.773856 4.8165112
#> 2   1.9356487  0.5839448  3.449522 0.9773000
#> 3   1.4813246  0.6924979  3.364823 1.2401766
#> 4   5.2589595  4.9973483  7.229810 4.2880013
#> 5   7.5048297  6.8603433  9.510839 5.7112143
#> 6   4.4967813  3.8647617  6.499266 2.8583244
#> 7   0.2845837  1.2038618  1.918305 2.5852961
#> 8   6.7661150  6.0826513  8.768831 4.9189344
#> 9   5.0235309  4.7550714  6.998316 4.0567141
#> 10  4.1452457  3.4194843  6.133517 2.3551698
#> 11  2.2166530  0.8742116  3.757006 0.7086590
#> 12  1.4564757  0.5982509  2.619917 1.8467783
#> 13  3.8114819  3.2455306  5.816352 2.3876067
#> 14  3.4559021  2.2950512  5.238077 0.8782851
#> 15 11.1695595 10.6646866 13.166827 9.5754736
#> 16  2.8133687  2.2476461  4.811931 1.6395710
#> 17  4.4180007  4.0299963  6.417378 3.2708872
#> 18  1.2992129  1.7747718  1.250715 3.1856669
#> 19  3.9190428  4.3631404  5.585831 4.4555823
#> 20  0.9505175  1.1199795  1.843919 2.5414470
#> 21  1.9309254  2.2416085  3.841368 2.6075533
#> 22  0.0000000  1.3521533  2.006836 2.6661800
#> 23  1.3521533  0.0000000  2.944807 1.4320021
#> 24  2.0068357  2.9448075  0.000000 4.3761390
#> 25  2.6661800  1.4320021  4.376139 0.0000000