Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1        2        3         4         5        6         7        8
#> 1  0.0000000 4.975053 2.056403 4.1514881 7.5264888 4.663123 6.4877117 2.254570
#> 2  4.9750532 0.000000 3.711151 1.5570709 3.9182818 2.186688 2.8570134 7.213233
#> 3  2.0564034 3.711151 0.000000 3.5491545 7.0957816 4.333255 5.9777458 4.126694
#> 4  4.1514881 1.557071 3.549155 0.0000000 3.5531537 0.874688 2.4435451 6.267120
#> 5  7.5264888 3.918282 7.095782 3.5531537 0.0000000 2.872045 1.1227082 9.435455
#> 6  4.6631227 2.186688 4.333255 0.8746880 2.8720449 0.000000 1.8290843 6.648175
#> 7  6.4877117 2.857013 5.977746 2.4435451 1.1227082 1.829084 0.0000000 8.468656
#> 8  2.2545701 7.213233 4.126694 6.2671203 9.4354545 6.648175 8.4686560 0.000000
#> 9  2.5705268 2.795967 2.523833 1.6211482 4.9714622 2.100191 3.9188832 4.646493
#> 10 0.8177079 4.208875 1.344500 3.5175240 6.9928603 4.122249 5.9199562 3.062396
#> 11 2.8856783 3.699593 3.529146 2.2319059 4.8991467 2.245151 3.9902490 4.551348
#> 12 4.8672156 1.140699 4.062898 0.7656643 3.0865154 1.074693 1.9666674 7.012923
#> 13 5.8065543 1.798476 5.066101 1.6553286 2.1406319 1.400044 1.0585551 7.902658
#> 14 7.4132890 3.362618 6.768169 3.3013305 0.8085695 2.780407 0.9796129 9.428548
#> 15 5.9510173 1.565994 5.057257 1.8293991 2.3699924 1.758077 1.3765694 8.093607
#> 16 6.6794323 2.978490 6.153914 2.6260878 0.9606236 2.021625 0.1927025 8.661185
#> 17 0.5041145 5.240223 2.539038 4.2895941 7.5536462 4.719327 6.5480283 1.987297
#> 18 0.9746991 4.129459 1.143823 3.5108923 7.0134443 4.148212 5.9307262 3.193635
#> 19 5.6199241 3.835011 5.774925 2.5665709 2.4626317 1.710673 2.0089680 7.271898
#> 20 0.7339364 5.415484 2.079675 4.7393143 8.1800580 5.308780 7.1219777 2.052033
#> 21 1.8654239 4.575207 3.191559 3.2840562 6.1821834 3.474229 5.2576601 3.273363
#> 22 3.3779508 2.073458 2.968162 0.7828607 4.2328490 1.383631 3.1490278 5.484343
#> 23 0.3810547 5.294666 2.434462 4.4018217 7.7110460 4.864111 6.6930511 1.918804
#> 24 0.9271395 4.114191 1.811449 3.2259838 6.6107009 3.742852 5.5645464 3.103583
#> 25 0.5857325 5.541008 2.594816 4.6561559 7.9588906 5.115037 6.9440932 1.675498
#>            9        10       11        12        13        14        15
#> 1  2.5705268 0.8177079 2.885678 4.8672156 5.8065543 7.4132890 5.9510173
#> 2  2.7959669 4.2088745 3.699593 1.1406992 1.7984764 3.3626185 1.5659940
#> 3  2.5238332 1.3445002 3.529146 4.0628980 5.0661010 6.7681686 5.0572569
#> 4  1.6211482 3.5175240 2.231906 0.7656643 1.6553286 3.3013305 1.8293991
#> 5  4.9714622 6.9928603 4.899147 3.0865154 2.1406319 0.8085695 2.3699924
#> 6  2.1001915 4.1222490 2.245151 1.0746931 1.4000444 2.7804070 1.7580766
#> 7  3.9188832 5.9199562 3.990249 1.9666674 1.0585551 0.9796129 1.3765694
#> 8  4.6464932 3.0623960 4.551348 7.0129230 7.9026584 9.4285479 8.0936071
#> 9  0.0000000 2.0346660 1.198147 2.3742593 3.2607895 4.8434010 3.4497780
#> 10 2.0346660 0.0000000 2.663712 4.1960768 5.1655155 6.8144180 5.2724851
#> 11 1.1981469 2.6637118 0.000000 2.9668570 3.6270275 4.9677949 3.9330794
#> 12 2.3742593 4.1960768 2.966857 0.0000000 1.0036727 2.7052706 1.0842411
#> 13 3.2607895 5.1655155 3.627028 1.0036727 0.0000000 1.7025323 0.4485321
#> 14 4.8434010 6.8144180 4.967795 2.7052706 1.7025323 0.0000000 1.7996161
#> 15 3.4497780 5.2724851 3.933079 1.0842411 0.4485321 1.7996161 0.0000000
#> 16 4.1102424 6.1090307 4.179894 2.1272069 1.1866169 0.7946371 1.4620873
#> 17 2.6730990 1.2392657 2.769768 5.0299528 5.9338824 7.4941948 6.1123106
#> 18 2.0811232 0.2020291 2.786828 4.1698875 5.1496596 6.8120384 5.2397738
#> 19 3.2943396 5.2672445 2.762690 2.6943141 2.4630798 2.8628565 2.9116117
#> 20 3.2086743 1.2306678 3.612538 5.4256828 6.3909061 8.0292310 6.5029616
#> 21 1.7802315 1.9782242 1.283609 4.0487310 4.8174764 6.2316354 5.0813274
#> 22 0.8392415 2.7721032 1.636898 1.5359189 2.4313668 4.0463701 2.6105999
#> 23 2.7954556 1.1893543 2.964854 5.1327243 6.0524647 7.6321156 6.2167243
#> 24 1.6457833 0.5838110 2.110319 3.9473570 4.8805997 6.4869756 5.0315873
#> 25 3.0492271 1.4029857 3.192078 5.3868684 6.3066829 7.8845289 6.4709046
#>           16        17        18       19        20       21        22
#> 1  6.6794323 0.5041145 0.9746991 5.619924 0.7339364 1.865424 3.3779508
#> 2  2.9784901 5.2402232 4.1294588 3.835011 5.4154839 4.575207 2.0734580
#> 3  6.1539141 2.5390384 1.1438231 5.774925 2.0796746 3.191559 2.9681619
#> 4  2.6260878 4.2895941 3.5108923 2.566571 4.7393143 3.284056 0.7828607
#> 5  0.9606236 7.5536462 7.0134443 2.462632 8.1800580 6.182183 4.2328490
#> 6  2.0216253 4.7193270 4.1482123 1.710673 5.3087795 3.474229 1.3836308
#> 7  0.1927025 6.5480283 5.9307262 2.008968 7.1219777 5.257660 3.1490278
#> 8  8.6611852 1.9872966 3.1936352 7.271898 2.0520325 3.273363 5.4843433
#> 9  4.1102424 2.6730990 2.0811232 3.294340 3.2086743 1.780231 0.8392415
#> 10 6.1090307 1.2392657 0.2020291 5.267244 1.2306678 1.978224 2.7721032
#> 11 4.1798939 2.7697684 2.7868280 2.762690 3.6125379 1.283609 1.6368983
#> 12 2.1272069 5.0299528 4.1698875 2.694314 5.4256828 4.048731 1.5359189
#> 13 1.1866169 5.9338824 5.1496596 2.463080 6.3909061 4.817476 2.4313668
#> 14 0.7946371 7.4941948 6.8120384 2.862857 8.0292310 6.231635 4.0463701
#> 15 1.4620873 6.1123106 5.2397738 2.911612 6.5029616 5.081327 2.6105999
#> 16 0.0000000 6.7406691 6.1184470 2.137409 7.3123777 5.448589 3.3375017
#> 17 6.7406691 0.0000000 1.4215521 5.531950 1.0631519 1.608657 3.5070156
#> 18 6.1184470 1.4215521 0.0000000 5.345145 1.2863554 2.160457 2.7824902
#> 19 2.1374085 5.5319504 5.3451446 0.000000 6.3330243 4.006722 2.8513498
#> 20 7.3123777 1.0631519 1.2863554 6.333024 0.0000000 2.590674 3.9828760
#> 21 5.4485892 1.6086575 2.1604574 4.006722 2.5906738 0.000000 2.5502908
#> 22 3.3375017 3.5070156 2.7824902 2.851350 3.9828760 2.550291 0.0000000
#> 23 6.8854830 0.2358189 1.3535531 5.723700 0.8305032 1.831238 3.6214281
#> 24 5.7559808 1.1266278 0.7538577 4.774177 1.5777155 1.409240 2.4510716
#> 25 7.1365818 0.4231733 1.5535406 5.953798 0.7868149 2.025704 3.8757054
#>           23        24        25
#> 1  0.3810547 0.9271395 0.5857325
#> 2  5.2946655 4.1141906 5.5410079
#> 3  2.4344616 1.8114485 2.5948162
#> 4  4.4018217 3.2259838 4.6561559
#> 5  7.7110460 6.6107009 7.9588906
#> 6  4.8641113 3.7428524 5.1150366
#> 7  6.6930511 5.5645464 6.9440932
#> 8  1.9188041 3.1035828 1.6754978
#> 9  2.7954556 1.6457833 3.0492271
#> 10 1.1893543 0.5838110 1.4029857
#> 11 2.9648543 2.1103194 3.1920780
#> 12 5.1327243 3.9473570 5.3868684
#> 13 6.0524647 4.8805997 6.3066829
#> 14 7.6321156 6.4869756 7.8845289
#> 15 6.2167243 5.0315873 6.4709046
#> 16 6.8854830 5.7559808 7.1365818
#> 17 0.2358189 1.1266278 0.4231733
#> 18 1.3535531 0.7538577 1.5535406
#> 19 5.7236995 4.7741771 5.9537985
#> 20 0.8305032 1.5777155 0.7868149
#> 21 1.8312382 1.4092401 2.0257044
#> 22 3.6214281 2.4510716 3.8757054
#> 23 0.0000000 1.1921604 0.2543413
#> 24 1.1921604 0.0000000 1.4446226
#> 25 0.2543413 1.4446226 0.0000000