Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2        3         4        5         6           7
#> 1  0.0000000 2.1457365 1.490340 2.8304859 3.587547 1.6553509  1.20388801
#> 2  2.1457365 0.0000000 1.869027 2.6736376 2.232196 2.9299864  3.33293201
#> 3  1.4903404 1.8690272 0.000000 3.8321004 3.988300 3.0914452  2.20602015
#> 4  2.8304859 2.6736376 3.832100 0.0000000 1.692532 1.8024905  3.70407268
#> 5  3.5875467 2.2321963 3.988300 1.6925319 0.000000 3.2338366  4.71389301
#> 6  1.6553509 2.9299864 3.091445 1.8024905 3.233837 0.0000000  2.07838794
#> 7  1.2038880 3.3329320 2.206020 3.7040727 4.713893 2.0783879  0.00000000
#> 8  4.3062971 2.5301575 4.397571 2.7215837 1.029117 4.1957198  5.48568868
#> 9  2.9791379 5.1245451 3.986919 4.9122212 6.228797 3.1097437  1.82819759
#> 10 2.8992421 0.8551521 2.715824 2.5997302 1.629232 3.3574710  4.10223500
#> 11 3.2442966 1.2049049 3.055947 2.7116342 1.516450 3.6239526  4.44800345
#> 12 0.4744570 2.2025765 1.116758 3.2631477 3.887831 2.1234143  1.19426132
#> 13 5.5834405 3.4442239 5.060710 4.8228181 3.188778 6.0019477  6.77664334
#> 14 0.9512643 1.3008981 1.613569 2.2235686 2.679959 1.7490133  2.14881248
#> 15 1.8671147 0.3387328 1.810510 2.4196942 2.187300 2.5914430  3.06642174
#> 16 4.4027924 2.6008508 4.469638 2.8320388 1.139509 4.3064287  5.58478282
#> 17 4.6422235 2.7053496 4.561016 3.2763482 1.589671 4.6795872  5.83881378
#> 18 1.5696732 1.4788984 2.381426 1.4521857 2.043515 1.5280879  2.67037969
#> 19 1.5454568 0.6631627 1.686772 2.3029775 2.317151 2.2877700  2.74838578
#> 20 0.8011544 2.1136899 2.136987 2.0607262 3.030461 0.9711107  1.70493329
#> 21 1.1620658 3.2827051 2.135663 3.7047582 4.691114 2.0996676  0.07251795
#> 22 2.4694132 2.4685694 3.508446 0.3638634 1.841411 1.4866235  3.34420099
#> 23 8.8364879 6.7680375 8.496392 7.3352758 5.683374 8.8933414 10.04000163
#> 24 5.1122520 3.3571931 5.225694 3.2669938 1.646325 4.8800609  6.28048647
#> 25 2.4545946 4.4769174 3.801969 3.7493601 5.200591 1.9764431  1.68967189
#>            8         9        10        11       12       13        14
#> 1  4.3062971  2.979138 2.8992421 3.2442966 0.474457 5.583440 0.9512643
#> 2  2.5301575  5.124545 0.8551521 1.2049049 2.202576 3.444224 1.3008981
#> 3  4.3975713  3.986919 2.7158242 3.0559469 1.116758 5.060710 1.6135688
#> 4  2.7215837  4.912221 2.5997302 2.7116342 3.263148 4.822818 2.2235686
#> 5  1.0291165  6.228797 1.6292317 1.5164502 3.887831 3.188778 2.6799595
#> 6  4.1957198  3.109744 3.3574710 3.6239526 2.123414 6.001948 1.7490133
#> 7  5.4856887  1.828198 4.1022350 4.4480035 1.194261 6.776643 2.1488125
#> 8  0.0000000  7.100357 1.7068665 1.4099252 4.531809 2.231595 3.3583660
#> 9  7.1003567  0.000000 5.8531772 6.1889468 3.022371 8.555695 3.8770780
#> 10 1.7068665  5.853177 0.0000000 0.3538465 3.017390 2.714815 1.9761056
#> 11 1.4099252  6.188947 0.3538465 0.0000000 3.370198 2.400371 2.3134575
#> 12 4.5318093  3.022371 3.0173903 3.3701981 0.000000 5.637486 1.2080525
#> 13 2.2315947  8.555695 2.7148149 2.4003714 5.637486 0.000000 4.6836059
#> 14 3.3583660  3.877078 1.9761056 2.3134575 1.208053 4.683606 0.0000000
#> 15 2.6214406  4.838093 1.0434911 1.3954926 1.975002 3.718050 0.9818841
#> 16 0.1118668  7.205553 1.7684364 1.4602587 4.622574 2.141837 3.4537790
#> 17 0.5764970  7.507514 1.8503083 1.5051939 4.822366 1.664904 3.6915307
#> 18 2.8620734  4.238619 1.8296441 2.1043817 1.930583 4.496980 0.7845264
#> 19 2.8542579  4.508502 1.3538750 1.7004385 1.689833 4.047353 0.6452040
#> 20 3.8637317  3.238926 2.6978549 3.0109030 1.268569 5.409442 0.8351038
#> 21 5.4538287  1.890450 4.0574888 4.4042420 1.132404 6.726864 2.1104791
#> 22 2.8569972  4.589700 2.5050406 2.6615133 2.905878 4.876974 1.8952053
#> 23 4.6976427 11.741801 5.9485571 5.5978249 8.963664 3.473422 7.8939303
#> 24 0.8302830  7.855615 2.5244613 2.2105471 5.351929 2.266183 4.1688959
#> 25 6.1381821  1.297360 5.0941779 5.4035459 2.696813 7.803633 3.1761949
#>           15        16        17        18        19        20         21
#> 1  1.8671147 4.4027924 4.6422235 1.5696732 1.5454568 0.8011544 1.16206577
#> 2  0.3387328 2.6008508 2.7053496 1.4788984 0.6631627 2.1136899 3.28270512
#> 3  1.8105096 4.4696384 4.5610158 2.3814256 1.6867723 2.1369871 2.13566344
#> 4  2.4196942 2.8320388 3.2763482 1.4521857 2.3029775 2.0607262 3.70475820
#> 5  2.1873002 1.1395090 1.5896713 2.0435151 2.3171509 3.0304610 4.69111370
#> 6  2.5914430 4.3064287 4.6795872 1.5280879 2.2877700 0.9711107 2.09966757
#> 7  3.0664217 5.5847828 5.8388138 2.6703797 2.7483858 1.7049333 0.07251795
#> 8  2.6214406 0.1118668 0.5764970 2.8620734 2.8542579 3.8637317 5.45382870
#> 9  4.8380929 7.2055528 7.5075137 4.2386192 4.5085024 3.2389259 1.89044996
#> 10 1.0434911 1.7684364 1.8503083 1.8296441 1.3538750 2.6978549 4.05748883
#> 11 1.3954926 1.4602587 1.5051939 2.1043817 1.7004385 3.0109030 4.40424199
#> 12 1.9750021 4.6225742 4.8223662 1.9305829 1.6898327 1.2685689 1.13240359
#> 13 3.7180504 2.1418367 1.6649043 4.4969800 4.0473532 5.4094418 6.72686414
#> 14 0.9818841 3.4537790 3.6915307 0.7845264 0.6452040 0.8351038 2.11047909
#> 15 0.0000000 2.7023416 2.8600205 1.1550674 0.3367340 1.7814858 3.01978867
#> 16 2.7023416 0.0000000 0.4791524 2.9669704 2.9415319 3.9681135 5.55230468
#> 17 2.8600205 0.4791524 0.0000000 3.2800479 3.1326894 4.2708369 5.80177874
#> 18 1.1550674 2.9669704 3.2800479 0.0000000 0.9289524 1.0023841 2.64804101
#> 19 0.3367340 2.9415319 3.1326894 0.9289524 0.0000000 1.4505781 2.70398215
#> 20 1.7814858 3.9681135 4.2708369 1.0023841 1.4505781 0.0000000 1.69204509
#> 21 3.0197887 5.5523047 5.8017787 2.6480410 2.7039822 1.6920451 0.00000000
#> 22 2.1896270 2.9687733 3.3907279 1.1436663 2.0362630 1.6969872 3.34385553
#> 23 6.9919986 4.5870671 4.2400102 7.5200495 7.2969924 8.5088172 9.99826984
#> 24 3.4515297 0.7567718 0.8575192 3.6304774 3.6803807 4.6317420 6.25127423
#> 25 4.1556008 6.2474521 6.5953980 3.3309633 3.8192818 2.3965417 1.76076382
#>           22        23        24        25
#> 1  2.4694132  8.836488 5.1122520  2.454595
#> 2  2.4685694  6.768037 3.3571931  4.476917
#> 3  3.5084463  8.496392 5.2256937  3.801969
#> 4  0.3638634  7.335276 3.2669938  3.749360
#> 5  1.8414110  5.683374 1.6463249  5.200591
#> 6  1.4866235  8.893341 4.8800609  1.976443
#> 7  3.3442010 10.040002 6.2804865  1.689672
#> 8  2.8569972  4.697643 0.8302830  6.138182
#> 9  4.5897000 11.741801 7.8556149  1.297360
#> 10 2.5050406  5.948557 2.5244613  5.094178
#> 11 2.6615133  5.597825 2.2105471  5.403546
#> 12 2.9058782  8.963664 5.3519294  2.696813
#> 13 4.8769738  3.473422 2.2661825  7.803633
#> 14 1.8952053  7.893930 4.1688959  3.176195
#> 15 2.1896270  6.991999 3.4515297  4.155601
#> 16 2.9687733  4.587067 0.7567718  6.247452
#> 17 3.3907279  4.240010 0.8575192  6.595398
#> 18 1.1436663  7.520049 3.6304774  3.330963
#> 19 2.0362630  7.296992 3.6803807  3.819282
#> 20 1.6969872  8.508817 4.6317420  2.396542
#> 21 3.3438555  9.998270 6.2512742  1.760764
#> 22 0.0000000  7.522148 3.4652872  3.454726
#> 23 7.5221485  0.000000 4.0687151 10.829752
#> 24 3.4652872  4.068715 0.0000000  6.846190
#> 25 3.4547260 10.829752 6.8461899  0.000000