Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3          4        5         6         7
#> 1  0.0000000 5.0156975 1.7362755  4.9465302 3.625946 1.0741615 0.2602028
#> 2  5.0156975 0.0000000 6.7487292  9.6134625 2.163089 6.0672108 4.8738154
#> 3  1.7362755 6.7487292 0.0000000  3.4770909 5.237635 0.6953203 1.8807816
#> 4  4.9465302 9.6134625 3.4770909  0.0000000 7.700098 3.9348781 4.9709504
#> 5  3.6259460 2.1630892 5.2376346  7.7000980 0.000000 4.5443897 3.4114303
#> 6  1.0741615 6.0672108 0.6953203  3.9348781 4.544390 0.0000000 1.1936289
#> 7  0.2602028 4.8738154 1.8807816  4.9709504 3.411430 1.1936289 0.0000000
#> 8  4.5240223 1.4011649 6.2338696  9.3941525 2.801399 5.5971089 4.4484552
#> 9  5.7953389 2.2704294 7.4603944 10.7183982 4.139099 6.8561801 5.7491788
#> 10 1.0114686 5.4397792 1.6203424  4.2213244 3.717016 1.0026976 0.8837124
#> 11 1.1475571 6.1195404 0.7919942  4.2644570 4.770389 0.5895587 1.3617388
#> 12 3.0448317 4.7535493 4.1306211  7.6065986 4.719375 3.7747273 3.1936861
#> 13 2.5292350 2.9397337 4.1298762  6.7016441 1.108295 3.4362972 2.3082252
#> 14 5.4614561 0.4985616 7.1892354  9.9841391 2.397922 6.5032047 5.3098911
#> 15 3.5699596 1.4929123 5.3060645  8.3141368 1.708522 4.6362041 3.4502745
#> 16 0.9407139 4.0782147 2.6769849  5.8059036 2.817376 2.0090993 0.8412803
#> 17 4.5958965 9.0534048 3.3315085  0.8166915 7.076180 3.6640073 4.5806231
#> 18 0.4123066 4.8595231 1.9192507  4.9208623 3.336980 1.2247607 0.1621539
#> 19 2.4567838 2.6541679 4.1554872  6.9703082 1.299179 3.4629902 2.2774114
#> 20 3.6935307 8.5045614 2.1662782  1.3121533 6.697056 2.6534010 3.7435549
#> 21 1.0722897 4.4292587 2.5847078  5.9592253 3.485553 2.0260069 1.1760139
#> 22 2.1559942 4.1695716 3.5193753  6.9655800 3.791496 3.0406885 2.2609175
#> 23 3.1310726 8.1405250 1.3949756  2.5265519 6.570996 2.0742453 3.2677650
#> 24 0.3880922 5.2207910 1.5523092  4.6128902 3.697511 0.8576853 0.3617375
#> 25 3.3863081 2.8503763 4.9775586  8.3308388 3.265405 4.4110791 3.3902590
#>           8         9        10        11       12        13        14       15
#> 1  4.524022  5.795339 1.0114686 1.1475571 3.044832 2.5292350 5.4614561 3.569960
#> 2  1.401165  2.270429 5.4397792 6.1195404 4.753549 2.9397337 0.4985616 1.492912
#> 3  6.233870  7.460394 1.6203424 0.7919942 4.130621 4.1298762 7.1892354 5.306065
#> 4  9.394152 10.718398 4.2213244 4.2644570 7.606599 6.7016441 9.9841391 8.314137
#> 5  2.801399  4.139099 3.7170155 4.7703885 4.719375 1.1082955 2.3979219 1.708522
#> 6  5.597109  6.856180 1.0026976 0.5895587 3.774727 3.4362972 6.5032047 4.636204
#> 7  4.448455  5.749179 0.8837124 1.3617388 3.193686 2.3082252 5.3098911 3.450275
#> 8  0.000000  1.415917 5.1769819 5.5183315 3.541370 3.1085596 1.8361952 1.217046
#> 9  1.415917  0.000000 6.5209344 6.7108170 4.200098 4.5234505 2.4824476 2.631757
#> 10 5.176982  6.520934 0.0000000 1.5019640 4.055155 2.6274801 5.8410102 4.096816
#> 11 5.518332  6.710817 1.5019640 0.0000000 3.350229 3.6693526 6.5760864 4.650286
#> 12 3.541370  4.200098 4.0551546 3.3502291 0.000000 4.0470867 5.2473395 3.444545
#> 13 3.108560  4.523450 2.6274801 3.6693526 4.047087 0.0000000 3.2852100 1.891732
#> 14 1.836195  2.482448 5.8410102 6.5760864 5.247340 3.2852100 0.0000000 1.977441
#> 15 1.217046  2.631757 4.0968160 4.6502860 3.444545 1.8917322 1.9774409 0.000000
#> 16 3.607277  4.912495 1.6370035 2.0507960 2.747480 1.7746284 4.5285591 2.629358
#> 17 8.929601 10.285491 3.7659155 4.0830735 7.424420 6.1193911 9.4029060 7.809849
#> 18 4.480781  5.798175 0.7707714 1.4561490 3.338478 2.2297660 5.2877125 3.454877
#> 19 2.599780  4.013360 2.7858663 3.5999856 3.518929 0.6082654 3.0600759 1.402535
#> 20 8.190456  9.486928 3.0655188 2.9554757 6.294675 5.6472926 8.9001099 7.151568
#> 21 3.706681  4.882434 2.0557843 1.8286266 2.033040 2.5150530 4.9083667 2.936844
#> 22 3.159188  4.132115 3.1426291 2.7297539 1.043713 3.0371501 4.6675967 2.734810
#> 23 7.619650  8.824139 2.8678577 2.1133220 5.277350 5.4690346 8.5774387 6.700985
#> 24 4.810178  6.109168 0.6336273 1.1292343 3.430173 2.5894303 5.6516963 3.806285
#> 25 1.629251  2.533459 4.2347274 4.2099642 1.921149 2.9815312 3.3389929 1.687853
#>           16         17        18        19       20        21       22
#> 1  0.9407139  4.5958965 0.4123066 2.4567838 3.693531 1.0722897 2.155994
#> 2  4.0782147  9.0534048 4.8595231 2.6541679 8.504561 4.4292587 4.169572
#> 3  2.6769849  3.3315085 1.9192507 4.1554872 2.166278 2.5847078 3.519375
#> 4  5.8059036  0.8166915 4.9208623 6.9703082 1.312153 5.9592253 6.965580
#> 5  2.8173757  7.0761801 3.3369795 1.2991788 6.697056 3.4855527 3.791496
#> 6  2.0090993  3.6640073 1.2247607 3.4629902 2.653401 2.0260069 3.040688
#> 7  0.8412803  4.5806231 0.1621539 2.2774114 3.743555 1.1760139 2.260917
#> 8  3.6072771  8.9296010 4.4807814 2.5997799 8.190456 3.7066813 3.159188
#> 9  4.9124947 10.2854913 5.7981754 4.0133603 9.486928 4.8824341 4.132115
#> 10 1.6370035  3.7659155 0.7707714 2.7858663 3.065519 2.0557843 3.142629
#> 11 2.0507960  4.0830735 1.4561490 3.5999856 2.955476 1.8286266 2.729754
#> 12 2.7474804  7.4244198 3.3384781 3.5189292 6.294675 2.0330399 1.043713
#> 13 1.7746284  6.1193911 2.2297660 0.6082654 5.647293 2.5150530 3.037150
#> 14 4.5285591  9.4029060 5.2877125 3.0600759 8.900110 4.9083667 4.667597
#> 15 2.6293584  7.8098493 3.4548771 1.4025345 7.151568 2.9368444 2.734810
#> 16 0.0000000  5.3935262 0.8873679 1.5726345 4.584619 0.8159561 1.721145
#> 17 5.3935262  0.0000000 4.5080302 6.4365439 1.355172 5.6483024 6.697145
#> 18 0.8873679  4.5080302 0.0000000 2.2395668 3.711227 1.3124545 2.391761
#> 19 1.5726345  6.4365439 2.2395668 0.0000000 5.850393 2.1907743 2.540708
#> 20 4.5846192  1.3551716 3.7112272 5.8503935 0.000000 4.6786002 5.667104
#> 21 0.8159561  5.6483024 1.3124545 2.1907743 4.678600 0.0000000 1.087454
#> 22 1.7211447  6.6971454 2.3917615 2.5407078 5.667104 1.0874544 0.000000
#> 23 4.0717745  2.6937622 3.2939395 5.5333752 1.338606 3.9418557 4.803019
#> 24 1.2030149  4.2350215 0.3670767 2.6061516 3.381996 1.4573138 2.542829
#> 25 2.6005487  7.9703588 3.4713676 2.3733647 7.063563 2.3944343 1.604794
#>          23        24       25
#> 1  3.131073 0.3880922 3.386308
#> 2  8.140525 5.2207910 2.850376
#> 3  1.394976 1.5523092 4.977559
#> 4  2.526552 4.6128902 8.330839
#> 5  6.570996 3.6975111 3.265405
#> 6  2.074245 0.8576853 4.411079
#> 7  3.267765 0.3617375 3.390259
#> 8  7.619650 4.8101784 1.629251
#> 9  8.824139 6.1091679 2.533459
#> 10 2.867858 0.6336273 4.234727
#> 11 2.113322 1.1292343 4.209964
#> 12 5.277350 3.4301729 1.921149
#> 13 5.469035 2.5894303 2.981531
#> 14 8.577439 5.6516963 3.338993
#> 15 6.700985 3.8062853 1.687853
#> 16 4.071774 1.2030149 2.600549
#> 17 2.693762 4.2350215 7.970359
#> 18 3.293940 0.3670767 3.471368
#> 19 5.533375 2.6061516 2.373365
#> 20 1.338606 3.3819956 7.063563
#> 21 3.941856 1.4573138 2.394434
#> 22 4.803019 2.5428294 1.604794
#> 23 0.000000 2.9278711 6.319181
#> 24 2.927871 0.0000000 3.736274
#> 25 6.319181 3.7362739 0.000000