Pairwise Distance Matrix

Given a set of locations, this function computes the distances between each possible pair of locations.

Usage

pdist(x, y, geo = FALSE)

Arguments

x

A vector of x (or longitude) coordinates.

y

A vector of y (or latitude) coordinates.

geo

A logical value indicating whether the locations are defined by geographic coordinates (pairs of longitude/latitude values). Default: FALSE.

Value

A square matrix representing pairwise distances between each possible pair of locations.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

x <- rnorm(25)
y <- rnorm(25, sd = 3)
pdist(x, y)
#>            1         2         3         4        5          6          7
#> 1  0.0000000 3.3712461  4.646710  4.267491 4.705421  5.1977378  5.0447697
#> 2  3.3712461 0.0000000  5.567724  4.453900 1.424320  2.3523915  2.0688972
#> 3  4.6467103 5.5677237  0.000000  1.311115 5.963916  7.9109390  7.6085692
#> 4  4.2674914 4.4539000  1.311115  0.000000 4.705625  6.7604008  6.4451971
#> 5  4.7054209 1.4243197  5.963916  4.705625 0.000000  2.4082234  2.0656693
#> 6  5.1977378 2.3523915  7.910939  6.760401 2.408223  0.0000000  0.3430821
#> 7  5.0447697 2.0688972  7.608569  6.445197 2.065669  0.3430821  0.0000000
#> 8  2.6637528 4.7677100  2.248719  2.521340 5.666335  7.0468258  6.8024344
#> 9  3.2668253 3.7366416  1.837809  1.006933 4.249316  6.0854031  5.7888155
#> 10 0.2009317 3.1841811  4.567499  4.139643 4.510371  5.0482342  4.8874221
#> 11 5.6849261 7.8456149  2.966672  4.218473 8.561059 10.1676745  9.9070352
#> 12 3.0516936 4.2389055  1.616371  1.438071 4.913641  6.5859160  6.3077953
#> 13 1.8116687 1.5596160  4.888418  4.040338 2.918399  3.5521129  3.3503125
#> 14 1.4053909 1.9809141  4.920746  4.173015 3.355738  3.8489613  3.6721091
#> 15 4.7094048 1.9908217  5.034148  3.737015 1.158805  3.5569254  3.2138434
#> 16 8.1273263 5.9110305 11.472130 10.358524 5.950424  3.6167235  3.9484286
#> 17 2.2367930 3.9958084  2.412689  2.254948 4.894695  6.2881181  6.0381352
#> 18 4.7286674 6.5904860  1.735452  2.925537 7.256013  8.9289902  8.6581861
#> 19 4.5702596 6.8060047  2.444727  3.531502 7.598014  9.1086692  8.8567952
#> 20 5.2471744 7.4925486  2.847831  4.041337 8.255905  9.8012846  9.5470253
#> 21 4.5569150 3.5585730  2.825026  1.523719 3.435413  5.6941808  5.3598992
#> 22 6.2844364 7.9589568  2.541560  3.850311 8.477217 10.3108216 10.0245408
#> 23 2.2562846 4.2111128  2.437011  2.406869 5.133887  6.4884174  6.2439315
#> 24 3.3932975 4.2435344  1.391033  1.005632 4.796831  6.5956891  6.3047224
#> 25 3.5365792 0.2096265  5.768114  4.644209 1.343593  2.1472596  1.8606915
#>             8         9        10         11         12        13        14
#> 1   2.6637528 3.2668253 0.2009317  5.6849261  3.0516936 1.8116687 1.4053909
#> 2   4.7677100 3.7366416 3.1841811  7.8456149  4.2389055 1.5596160 1.9809141
#> 3   2.2487194 1.8378086 4.5674985  2.9666719  1.6163711 4.8884177 4.9207465
#> 4   2.5213402 1.0069325 4.1396430  4.2184731  1.4380708 4.0403383 4.1730153
#> 5   5.6663347 4.2493158 4.5103708  8.5610588  4.9136405 2.9183993 3.3557377
#> 6   7.0468258 6.0854031 5.0482342 10.1676745  6.5859160 3.5521129 3.8489613
#> 7   6.8024344 5.7888155 4.8874221  9.9070352  6.3077953 3.3503125 3.6721091
#> 8   0.0000000 1.8859157 2.6467045  3.1672558  1.1239705 3.5656777 3.4136387
#> 9   1.8859157 0.0000000 3.1351617  4.3320602  0.7958489 3.1086697 3.2027276
#> 10  2.6467045 3.1351617 0.0000000  5.7129300  2.9615761 1.6257242 1.2301108
#> 11  3.1672558 4.3320602 5.7129300  0.0000000  3.6514495 6.7293718 6.5785374
#> 12  1.1239705 0.7958489 2.9615761  3.6514495  0.0000000 3.3447603 3.3341019
#> 13  3.5656777 3.1086697 1.6257242  6.7293718  3.3447603 0.0000000 0.4410478
#> 14  3.4136387 3.2027276 1.2301108  6.5785374  3.3341019 0.4410478 0.0000000
#> 15  5.0962143 3.4546876 4.5084824  7.7801406  4.1959901 3.0971199 3.5267184
#> 16 10.3777740 9.6356667 8.0286156 13.5429635 10.0680925 6.8136478 6.9869559
#> 17  0.7769474 1.3930011 2.1685111  3.8918677  0.8601464 2.8469788 2.7412815
#> 18  2.0795367 3.0169843 4.7222432  1.3194387  2.3621634 5.5616798 5.4564787
#> 19  2.0733501 3.4410045 4.5969524  1.1160413  2.6974959 5.6387341 5.4743712
#> 20  2.7710875 4.0591205 5.2802172  0.4569464  3.3434554 6.3366583 6.1717175
#> 21  3.6434346 1.7581795 4.3838820  5.7381712  2.5352330 3.6936299 3.9729298
#> 22  3.6287316 4.2539606 6.2748800  1.2611569  3.7364789 7.0414767 6.9662871
#> 23  0.5585536 1.5967931 2.2094902  3.7134899  0.9762713 3.0158829 2.8812823
#> 24  1.5390500 0.5494166 3.2879591  3.7834056  0.4326168 3.4973853 3.5374209
#> 25  4.9761804 3.9391241 3.3524203  8.0552189  4.4472417 1.7269178 2.1376883
#>          15        16        17         18         19         20       21
#> 1  4.709405  8.127326 2.2367930  4.7286674  4.5702596  5.2471744 4.556915
#> 2  1.990822  5.911031 3.9958084  6.5904860  6.8060047  7.4925486 3.558573
#> 3  5.034148 11.472130 2.4126889  1.7354523  2.4447275  2.8478307 2.825026
#> 4  3.737015 10.358524 2.2549479  2.9255368  3.5315016  4.0413369 1.523719
#> 5  1.158805  5.950424 4.8946948  7.2560128  7.5980144  8.2559049 3.435413
#> 6  3.556925  3.616724 6.2881181  8.9289902  9.1086692  9.8012846 5.694181
#> 7  3.213843  3.948429 6.0381352  8.6581861  8.8567952  9.5470253 5.359899
#> 8  5.096214 10.377774 0.7769474  2.0795367  2.0733501  2.7710875 3.643435
#> 9  3.454688  9.635667 1.3930011  3.0169843  3.4410045  4.0591205 1.758180
#> 10 4.508482  8.028616 2.1685111  4.7222432  4.5969524  5.2802172 4.383882
#> 11 7.780141 13.542964 3.8918677  1.3194387  1.1160413  0.4569464 5.738171
#> 12 4.195990 10.068093 0.8601464  2.3621634  2.6974959  3.3434554 2.535233
#> 13 3.097120  6.813648 2.8469788  5.5616798  5.6387341  6.3366583 3.693630
#> 14 3.526718  6.986956 2.7412815  5.4564787  5.4743712  6.1717175 3.972930
#> 15 0.000000  7.108717 4.3591308  6.4614685  6.8885012  7.5137892 2.353330
#> 16 7.108717  0.000000 9.6574489 12.3615972 12.4493692 13.1474746 9.308141
#> 17 4.359131  9.657449 0.0000000  2.7190077  2.8205953  3.5142929 3.107785
#> 18 6.461468 12.361597 2.7190077  0.0000000  0.7810377  1.1196962 4.436491
#> 19 6.888501 12.449369 2.8205953  0.7810377  0.0000000  0.6981054 5.003189
#> 20 7.513789 13.147475 3.5142929  1.1196962  0.6981054  0.0000000 5.545767
#> 21 2.353330  9.308141 3.1077847  4.4364906  5.0031886  5.5457672 0.000000
#> 22 7.573436 13.803939 4.2263469  1.5567881  1.9534848  1.5885013 5.348682
#> 23 4.607918  9.829488 0.2493829  2.5792019  2.6274858  3.3243319 3.331268
#> 24 3.999969 10.124658 1.2668022  2.4676963  2.9078032  3.5156220 2.174819
#> 25 2.031819  5.716203 4.2046798  6.7995941  7.0154629  7.7021213 3.714240
#>           22        23         24        25
#> 1   6.284436 2.2562846  3.3932975 3.5365792
#> 2   7.958957 4.2111128  4.2435344 0.2096265
#> 3   2.541560 2.4370108  1.3910326 5.7681140
#> 4   3.850311 2.4068694  1.0056322 4.6442088
#> 5   8.477217 5.1338868  4.7968308 1.3435928
#> 6  10.310822 6.4884174  6.5956891 2.1472596
#> 7  10.024541 6.2439315  6.3047224 1.8606915
#> 8   3.628732 0.5585536  1.5390500 4.9761804
#> 9   4.253961 1.5967931  0.5494166 3.9391241
#> 10  6.274880 2.2094902  3.2879591 3.3524203
#> 11  1.261157 3.7134899  3.7834056 8.0552189
#> 12  3.736479 0.9762713  0.4326168 4.4472417
#> 13  7.041477 3.0158829  3.4973853 1.7269178
#> 14  6.966287 2.8812823  3.5374209 2.1376883
#> 15  7.573436 4.6079184  3.9999691 2.0318189
#> 16 13.803939 9.8294884 10.1246575 5.7162034
#> 17  4.226347 0.2493829  1.2668022 4.2046798
#> 18  1.556788 2.5792019  2.4676963 6.7995941
#> 19  1.953485 2.6274858  2.9078032 7.0154629
#> 20  1.588501 3.3243319  3.5156220 7.7021213
#> 21  5.348682 3.3312684  2.1748186 3.7142398
#> 22  0.000000 4.1084243  3.7233451 8.1654993
#> 23  4.108424 0.0000000  1.4037986 4.4193557
#> 24  3.723345 1.4037986  0.0000000 4.4485271
#> 25  8.165499 4.4193557  4.4485271 0.0000000